% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_fictive_fatal_prod_demand}
\alias{add_fictive_fatal_prod_demand}
\title{Add fictive production and fictive load to avoid infeasabilities with binding constraints}
\usage{
add_fictive_fatal_prod_demand(area, opts = antaresRead::simOptions())
}
\arguments{
\item{area}{A valid Antares area.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Add fictive production and fictive load to avoid infeasabilities with binding constraints
}
\keyword{internal}
