% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBand.R
\name{addBand}
\alias{addBand}
\title{Add Band to Water Values}
\usage{
addBand(vu, states, bandwidth = 100, failure_cost = 3000)
}
\arguments{
\item{vu}{Water values from \code{meanGridLayer}}

\item{states}{Level of states}

\item{bandwidth}{Ith of the band to add.}

\item{failure_cost}{Cost of failure}
}
\value{
A vector of water values with a band
}
\description{
Add Band to Water Values
}
\examples{
\dontrun{

# TODO

}
}
