% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBand.R
\name{addBandVU}
\alias{addBandVU}
\title{Add Band to Value Nodes}
\usage{
addBandVU(value_nodes, type = c("linear", "spline"), bandwidth = 100,
  failure_cost = NULL, area = NULL, opts = antaresRead::simOptions())
}
\arguments{
\item{value_nodes}{\code{data.table} in output of \code{\link{meanGridLayer}}.}

\item{type}{Perform a linear or a spline interpolation.}

\item{bandwidth}{Ith of the band to add.}

\item{failure_cost}{Cost of failure, if not provided will retrieve 
failure cost in Antares, for that you should specify a valid area.}

\item{area}{An Antares area.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
a \code{data.table} with one extra column
}
\description{
Add Band to Value Nodes
}
\examples{
\dontrun{

value_nodes <- addBandVU(value_nodes, 
                         type = "linear", 
                         bandwidth = 100,
                         failure_cost = 3000)

}
}
