% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{adjust_wv}
\alias{adjust_wv}
\title{Adjust  water Values}
\usage{
adjust_wv(results_dt, value = 0)
}
\arguments{
\item{results_dt}{Output from \code{watervalues} or \code{Grid_Matrix}}

\item{value}{value to add to water values}
}
\value{
a \code{data.table}
}
\description{
Adjust  water Values
}
