% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bellman.R
\name{Bellman}
\alias{Bellman}
\title{Compute Bellman values at step i from step i+1}
\usage{
Bellman(
  Data_week,
  next_week_values_l,
  decision_space,
  E_max,
  niveau_max,
  method,
  na_rm = TRUE,
  max_mcyear,
  print_test = FALSE,
  correct_outliers = FALSE,
  q_ratio = 0.75,
  test_week,
  counter,
  inaccessible_states = F,
  ...
)
}
\arguments{
\item{Data_week}{A "data.table" generated in Grid_Matrix code
that contains:
 * states Numeric. All the water values that can be set, listed in
  decreasing order.
 * value_inflow Numeric. Inflow values for each Monte-Carlo year.
 * Rewards for each simulation value and each Monte-Carlo year.
 * level_high Numeric. Highest possible reservoir value.
 * level_low Numeric. Lowest possible reservoir value.
 * states_next List of vectors enumerating all reachable states}

\item{next_week_values_l}{Numeric. Bellman values at step i+1.}

\item{decision_space}{Simulation constraints values}

\item{E_max}{Numeric of length 1. Maximum energy that can be generated by
hydro storage over one step of time.}

\item{niveau_max}{Numeric of length 1. Reservoir capacity.}

\item{method}{Character. Perform mean grid algorithm or grid mean algorithm?}

\item{na_rm}{Boolean. Remove NAs}

\item{max_mcyear}{Numeric of length 1. Number of Monte-Carlo year
used in simulations}

\item{print_test}{Boolean. print Bellman values.}

\item{correct_outliers}{If TRUE, outliers in Bellman values are replaced by spline
interpolations. Defaults to FALSE.}

\item{test_week}{Numeric of length 1. number of the week to print in test.}

\item{inaccessible_states}{Boolean. True to delete unaccessible states of any scenario in the result.}
}
\value{
a \code{data.table} like Data_week with the Bellman values
}
\description{
Compute Bellman values at step i from step i+1
}
