% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bellman.R
\name{bellman_parallel_value}
\alias{bellman_parallel_value}
\title{Compute Bellman values at step i from step i+1 (version used in parallel computing)}
\usage{
bellman_parallel_value(
  Data_week,
  i,
  next_week_values_l,
  max_mcyear,
  E_max,
  niveau_max,
  na_rm,
  decision_space,
  alpha,
  decimals
)
}
\arguments{
\item{Data_week}{A "data.table" generated in Grid_Matrix code
that contains:
 * states Numeric. All the water values that can be set, listed in
  decreasing order.
 * value_inflow Numeric. Inflow values for each Monte-Carlo year.
 * Rewards for each simulation value and each Monte-Carlo year.
 * level_high Numeric. Highest possible reservoir value.
 * level_low Numeric. Lowest possible reservoir value.
 * states_next List of vectors enumerating all reachable states}

\item{i}{the row number of the point to calculate her bellman value.}

\item{next_week_values_l}{Numeric. Bellman values at step i+1.}

\item{max_mcyear}{Numeric of length 1. Number of Monte-Carlo year
used in simulations}

\item{E_max}{Numeric of length 1. Maximum energy that can be generated by
hydro storage over one step of time.}

\item{niveau_max}{Numeric of length 1. Reservoir capacity.}

\item{na_rm}{Boolean. Remove NAs}

\item{decision_space}{Simulation constraints values}

\item{alpha}{maximum tolerance.}

\item{decimals}{number of decimals.}
}
\value{
a \code{data.table} like Data_week with the Bellman values
}
\description{
Compute Bellman values at step i from step i+1 (version used in parallel computing)
}
