% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bellman.R
\name{Bellman_value}
\alias{Bellman_value}
\title{Compute Bellman values at step i from step i+1}
\usage{
Bellman_value(
  states,
  statesID,
  level_high,
  level_low,
  value_inflow,
  reward,
  states_next,
  next_week_values_l,
  decision_space,
  E_max,
  niveau_max,
  method,
  na_rm = TRUE,
  max_mcyear,
  print_test = FALSE,
  correct_outliers = FALSE,
  q_ratio = 0.75,
  test_week,
  counter,
  ...
)
}
\arguments{
\item{states}{Numeric. All the water values that can be set, listed in
decreasing order.}

\item{level_high}{Numeric. Highest possible reservoir value.}

\item{level_low}{Numeric. Lowest possible reservoir value.}

\item{value_inflow}{Numeric. Inflow values for each Monte-Carlo year.}

\item{states_next}{List of vectors enumerating all reachable states}

\item{next_week_values_l}{Numeric. Bellman values at step i+1.}

\item{decision_space}{Simulation constraints values}

\item{E_max}{Numeric of length 1. Maximum energy that can be generated by
hydro storage over one step of time.}

\item{niveau_max}{Numeric of length 1. Reservoir capacity.}

\item{method}{Character. Perform mean grid algorithm or grid mean algorithm?}

\item{na_rm}{Boolean. Remove NAs}

\item{max_mcyear}{Numeric of length 1. Number of Monte-Carlo year
used in simulations}

\item{print_test}{Boolean. print Bellman values.}

\item{correct_outliers}{If TRUE, outliers in Bellman values are replaced by spline
interpolations. Defaults to FALSE.}

\item{test_week}{Numeric of length 1. number of the week to print in test.}

\item{Rewards}{for each simulation value and each Monte-Carlo year.}
}
\value{
a \code{Numeric} the Bellman values
}
\description{
Compute Bellman values at step i from step i+1
}
