% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bellman_subfunctions.R
\name{build_all_possible_decisions}
\alias{build_all_possible_decisions}
\title{Create a data frame with all possible transition and associated next state for
each scenario and each state for one particular week, used in \code{Bellman}.
For each next state, Bellman value is calculated.}
\usage{
build_all_possible_decisions(
  Data_week,
  decision_space,
  f_next_value,
  mcyears,
  lvl_high,
  lvl_low,
  E_max,
  P_max,
  next_week_values,
  niveau_max,
  overflow_cost
)
}
\arguments{
\item{Data_week}{Data frame generated in \code{Grid_Matrix} code containing
list of states and years evaluated (we suppose there is only one week at a time)
as weel as inflows}

\item{decision_space}{Simulation constraints values for the week
generated by \code{get_Reward} or \code{runWaterValuesSimulation}}

\item{f_next_value}{List of approximated Bellman functions for the next week,
generated by \code{get_bellman_values_interpolation}}

\item{mcyears}{Vector of monte carlo years to take into account}

\item{lvl_high}{Double, top rule curve for the end of the week in MWh}

\item{lvl_low}{Double, bottom rule curve for the end of the week in MWh}

\item{E_max}{Double, weekly maximum turbining power for the week}

\item{P_max}{Double, weekly maximum pumping power for the week}

\item{next_week_values}{Vector of Bellman values for the next week}

\item{niveau_max}{Reservoir capacity in MWh}

\item{overflow_cost}{Cost for overflow (equal to spillage cost of the area)}
}
\value{
Data frame with same format as Data week with 3 more columns :
next_state, control (transition to next state) and
next value (Bellman value of next state)
}
\description{
Create a data frame with all possible transition and associated next state for
each scenario and each state for one particular week, used in \code{Bellman}.
For each next state, Bellman value is calculated.
}
\keyword{internal}
