% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanGridLayer.R
\name{calculate_value_node}
\alias{calculate_value_node}
\title{Compute Bellman values at step i from step i+1}
\usage{
calculate_value_node(states, states_next, value_reward, value_inflow,
  decision_space, level_high, level_low, value_node_next_week,
  niveau_max = 10, E_max = 1.344, method, na_rm = FALSE)
}
\arguments{
\item{states}{Numeric. All the water values that can be set, listed in
decreasing order.}

\item{states_next}{List of vectors enumerating all reachable states?}

\item{value_reward}{A data.table. Rewards for each simulation value and each
Monte-Carlo year.}

\item{value_inflow}{Numeric. Inflow values for each Monte-Carlo year.}

\item{decision_space}{Simulation values?}

\item{level_high}{Numeric. Highest reservoir value?}

\item{level_low}{Numeric. Lowest reservoir value?}

\item{value_node_next_week}{Numeric. Bellman values at
step i+1.}

\item{niveau_max}{Numeric of length 1. Reservoir capacity.}

\item{E_max}{Numeric of length 1. Maximum energy that can be generated by
hydro storage over one step of time?}

\item{method}{Character. Perform mean grid algorithm or grid mean algorithm?}

\item{na_rm}{Boolean. Remove NAs}
}
\description{
Compute Bellman values at step i from step i+1
}
