% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistock_sequential_plaia.R
\name{calculateRewardsSimulationsWithPlaia}
\alias{calculateRewardsSimulationsWithPlaia}
\title{Compute reward function for \code{node} with plaia implementation.
The reward function of each week and each year is evaluated on \code{nb_simulations} controls.
Called for each area of \code{getBellmanValuesSequentialMultiStockWithPlaia()}.
If they are several areas, the trajectories of the other storage are fixed on their \code{optimal_trend}.}
\usage{
calculateRewardsSimulationsWithPlaia(
  node,
  list_areas,
  list_efficiency,
  opts,
  mcyears,
  nb_simulations,
  optimal_traj,
  list_max_hydro_weekly
)
}
\arguments{
\item{node}{Character. Name of the area where the reward is computed}

\item{list_areas}{Vector of areas concerned by simulations.}

\item{list_efficiency}{List of double. Efficiency for each area generated by \code{getPumpEfficiency()}.}

\item{opts}{List of study parameters returned by the function \code{antaresRead::setSimulationPath(simulation="input")} in input mode.}

\item{mcyears}{Vector of integer. Monte Carlo years used to compute water values.}

\item{nb_simulations}{Number of controls to simulate}

\item{optimal_traj}{Data frame containing optimal trajectory for all areas}

\item{list_max_hydro_weekly}{List of data.frame. Generated by \code{get_max_hydro()} for each area.}
}
\value{
a \code{data_frame} containing the rewards returned by the function \code{get_Reward()}
}
\description{
Compute reward function for \code{node} with plaia implementation.
The reward function of each week and each year is evaluated on \code{nb_simulations} controls.
Called for each area of \code{getBellmanValuesSequentialMultiStockWithPlaia()}.
If they are several areas, the trajectories of the other storage are fixed on their \code{optimal_trend}.
}
