% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{check_Bellman_inc}
\alias{check_Bellman_inc}
\title{Check the monotonicity of the calculated Bellman values and return a success rate.}
\usage{
check_Bellman_inc(results)
}
\arguments{
\item{results}{A data.table contains the Bellman values
Obtained using the function \code{Grid_Matrix()}}
}
\description{
Check the monotonicity of the calculated Bellman values and return a success rate.
}
