% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{check_resh_vu_dec}
\alias{check_resh_vu_dec}
\title{Check the monotonicity of the calculated the water values after putting them
in Antares Format and return a success rate.}
\usage{
check_resh_vu_dec(reshaped_results)
}
\arguments{
\item{reshaped_results}{A matrix contains the water values
Obtained using the function \code{to_Antares_Format()}}
}
\description{
Check the monotonicity of the calculated the water values after putting them
in Antares Format and return a success rate.
}
