% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{check_vu_dec}
\alias{check_vu_dec}
\title{Check the monotonicity of the calculated  water values and return a success rate.}
\usage{
check_vu_dec(results)
}
\arguments{
\item{results}{A data.table contains the water values
Obtained using the function \code{Grid_Matrix()}}
}
\description{
Check the monotonicity of the calculated  water values and return a success rate.
}
