% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{constant_fill}
\alias{constant_fill}
\title{Fill constant water values}
\usage{
constant_fill(results_dt, max_wv, min_wv)
}
\arguments{
\item{results_dt}{Output from \code{watervalues} or \code{Grid_Matrix}}

\item{max_wv}{maximal water value to use}

\item{min_wv}{minimal water value to use}
}
\value{
a \code{data.table}
}
\description{
Fill constant water values
}
