% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_functions.R
\name{constraint_generator}
\alias{constraint_generator}
\title{Generate a list of control values, eg difference between final and initial level for each week.
Used to run simulations in \code{runWaterValuesSimulation} and also to estimate reward functions
in functions such as \code{Grid_Matrix} and \code{get_Reward}}
\usage{
constraint_generator(
  area,
  nb_disc_stock,
  pumping = F,
  pumping_efficiency = NULL,
  opts,
  max_hydro = NULL,
  inflow = NULL,
  mcyears = NULL
)
}
\arguments{
\item{area}{The area concerned by the simulation.}

\item{nb_disc_stock}{Number of constraint values wanted for each week}

\item{pumping}{Boolean. True to take into account the pumping.}

\item{pumping_efficiency}{between 0 and 1. the pumping efficiency ratio.}

\item{opts}{List of simulation parameters returned by the function \code{antaresRead::setSimulationPath}}

\item{max_hydro}{weekly pumping and turbining maximum powers, generated by \code{get_max_hydro}}

\item{inflow}{weekly inflow}

\item{mcyears}{Vector of years used to evaluate inflow}
}
\description{
Generate a list of control values, eg difference between final and initial level for each week.
Used to run simulations in \code{runWaterValuesSimulation} and also to estimate reward functions
in functions such as \code{Grid_Matrix} and \code{get_Reward}
}
