% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_functions.R
\name{constraint_week}
\alias{constraint_week}
\title{Generate constraint values for a week at a time used in \code{constraint_generator}
depending on the maximum powers for the week}
\usage{
constraint_week(
  pumping,
  efficiency,
  nb_disc_stock,
  reservoir_capacity,
  hydro,
  week
)
}
\arguments{
\item{pumping}{Boolean. True to take into account the pumping.}

\item{efficiency}{between 0 and 1. the pumping efficiency ratio.}

\item{nb_disc_stock}{Number of constraint values wanted for each week}

\item{reservoir_capacity}{Double, reservoir capacity}

\item{hydro}{Pumping and turbining maximum powers for the week, generated by \code{get_max_hydro}}

\item{week}{Current week for which to compute controls}
}
\value{
List of constraint values for the week
}
\description{
Generate constraint values for a week at a time used in \code{constraint_generator}
depending on the maximum powers for the week
}
\keyword{internal}
