% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{converged}
\alias{converged}
\title{test a difference vector convergence, used in \code{Grid_Matrix}}
\usage{
converged(diff_vect, conv = 1)
}
\arguments{
\item{diff_vect}{is a vector of water values differences}

\item{conv}{is the value from which the difference become converged}
}
\description{
test a difference vector convergence, used in \code{Grid_Matrix}
}
