% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{correct_concavity}
\alias{correct_concavity}
\title{Correct concavity of Bellman values to have nice monotony for water values}
\usage{
correct_concavity(df_value_node, weeks)
}
\arguments{
\item{df_value_node}{DataFrame containing bellman values}

\item{weeks}{Weeks for which we want to correct concavity}
}
\value{
vector of corrected Bellman values
}
\description{
Correct concavity of Bellman values to have nice monotony for water values
}
