% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{correct_outliers}
\alias{correct_outliers}
\title{Replace outliers values by spline}
\usage{
correct_outliers(vector)
}
\arguments{
\item{vector}{numeric vector to remove outliers values from it.}
}
\description{
Replace outliers values by spline
}
