% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{decr}
\alias{decr}
\title{Check if the vector has decreasing values}
\usage{
decr(vector)
}
\arguments{
\item{vector}{numeric vector whose decreasing check is wanted.}
}
\description{
Check if the vector has decreasing values
}
