% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_functions.R
\name{disable_constraint}
\alias{disable_constraint}
\title{This function disable binding constraints for \code{runWaterValuesSimulation}}
\usage{
disable_constraint(constraint_value, name_bc, pumping = F, opts, area = NULL)
}
\arguments{
\item{constraint_value}{the value of the constraint}

\item{name_bc}{the name of the constraint.}

\item{pumping}{Boolean. True to take into account the pumping.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{area}{Area used to calculate watervalues}
}
\description{
This function disable binding constraints for \code{runWaterValuesSimulation}
}
