% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optim_H2_multi_nodes_functions.R
\name{edit_study_with_results}
\alias{edit_study_with_results}
\title{Edit the study with the result from investment}
\usage{
edit_study_with_results(opts, node, output_node, candidates_types)
}
\arguments{
\item{opts}{List of study parameters returned by the function \code{antaresRead::setSimulationPath(simulation="input")} in input mode.}

\item{node}{Character. Area for which to update data.}

\item{output_node}{List. Output of \code{MultiStock_H2_Investment_reward_compute_once}.}

\item{candidates_types}{Data_frame with column names : c("index", "name", "type", "TOTEX", "Marg_price").
It is a parameter of \code{MultiStock_H2_Investment_reward_compute_once}.}
}
\description{
Edit the study with the result from investment
}
