% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_aggregated_results}
\alias{example_aggregated_results}
\title{Example of output of \code{Grid_Matrix}}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2704 rows and 13 columns.
}
\source{
Generated with a Antares toy study
}
\usage{
example_aggregated_results
}
\description{
A list containing a watervalues data frame and an aggregated_results data frame
}
\keyword{datasets}
