% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_results}
\alias{example_results}
\title{Example of output of \code{Grid_Matrix}}
\format{
An object of class \code{detailled and aggregated results of watervalues calculation} of length 2.
}
\source{
Generated with a Antares toy study
}
\usage{
example_results
}
\description{
A list containing a watervalues data frame and an aggregated_results data frame
}
\keyword{datasets}
