% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_functions.R
\name{generate_constraints}
\alias{generate_constraints}
\title{This function generate binding constraints for \code{runWaterValuesSimulation}}
\usage{
generate_constraints(pumping, efficiency, opts, area)
}
\arguments{
\item{pumping}{bool}

\item{efficiency}{in [0,1]. efficient ratio of pumping.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{area}{Area used to calculate watervalues}
}
\description{
This function generate binding constraints for \code{runWaterValuesSimulation}
}
