% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_functions.R
\name{generate_link_coeff}
\alias{generate_link_coeff}
\title{Generate coefficients for biding constraints, used in \code{runWaterValuesSimulation}}
\usage{
generate_link_coeff(area, fictive_area)
}
\arguments{
\item{area}{Area with the area}

\item{fictive_area}{Fictive area involved in the biding contraint}
}
\value{
Named vector of coefficients
}
\description{
Generate coefficients for biding constraints, used in \code{runWaterValuesSimulation}
}
