% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bellman_subfunctions.R
\name{get_bellman_values_interpolation}
\alias{get_bellman_values_interpolation}
\title{Create approximation of Bellman function for next week for each scenario, used in \code{Bellman} and
in \code{getOptimalTrend}}
\usage{
get_bellman_values_interpolation(next_state, next_week_values, mcyears)
}
\arguments{
\item{next_state}{possible states for next week}

\item{next_week_values}{Vector of Bellman values to interpolate}

\item{mcyears}{Vector of monte carlo years to take into account}
}
\value{
List of \code{stats::approxfun} for each scenario
}
\description{
Create approximation of Bellman function for next week for each scenario, used in \code{Bellman} and
in \code{getOptimalTrend}
}
\keyword{internal}
