% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_inflow}
\alias{get_inflow}
\title{Get inflow in a week, used in different functions}
\usage{
get_inflow(area, opts = antaresRead::simOptions(), mcyears)
}
\arguments{
\item{area}{The area concerned by the simulation.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{mcyears}{Vector of years used to evaluate cost}
}
\description{
Get inflow in a week, used in different functions
}
