% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readReservoirLevels.R
\name{get_initial_level_year_per_year}
\alias{get_initial_level_year_per_year}
\title{Get initial level year per year}
\usage{
get_initial_level_year_per_year(area, opts)
}
\arguments{
\item{area}{Character. The Antares area concerned by water values computation.}

\item{opts}{List of study parameters returned by the function \code{antaresRead::setSimulationPath(simulation="input")} in input mode.}
}
\value{
Vector. Initial level for each MC year in percentage, value between 0 and 100\%.
}
\description{
Get initial level of hydro storage for the given area for each MC year based on the scenario builder. If random in the scenario builder, the initial level is defined trough low and high reservoir levels for the first day.
}
