% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Reward.R
\name{get_local_reward}
\alias{get_local_reward}
\title{Calculate rewards for a simulation based on marginal prices}
\usage{
get_local_reward(
  opts,
  hours,
  possible_controls,
  T_max,
  P_max,
  area_price,
  mcyears,
  district_balance = "water values district",
  pump_eff = 1
)
}
\arguments{
\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{hours}{vector of hours used to evaluate costs/rewards of pumping/generating}

\item{possible_controls}{vector of controls evaluated}

\item{T_max}{Max generating power}

\item{P_max}{Max pumping power}

\item{area_price}{Area used to evaluate marginal prices}

\item{mcyears}{Vector of years used to evaluate rewards}

\item{district_balance}{Name of district used to evaluate controls on the stock}

\item{pump_eff}{Pumping efficiency}
}
\value{
a data.table {mcYear,week,u,reward}
}
\description{
Calculate rewards for a simulation based on marginal prices
}
