% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Reward.R
\name{get_local_reward_turb}
\alias{get_local_reward_turb}
\title{Calculate rewards for a simulation based on marginal prices for a stock with only generating power}
\usage{
get_local_reward_turb(
  opts,
  possible_controls,
  T_max,
  area_price,
  mcyears,
  district_balance = "water values district"
)
}
\arguments{
\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{possible_controls}{vector of controls evaluated}

\item{T_max}{Max generating power}

\item{area_price}{Area used to evaluate marginal prices}

\item{mcyears}{Vector of years used to evaluate rewards}

\item{district_balance}{Name of district used to evaluate controls on the stock}
}
\value{
a data.table {mcYear,week,u,reward}
}
\description{
Calculate rewards for a simulation based on marginal prices for a stock with only generating power
}
