% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_reservoir_capacity}
\alias{get_reservoir_capacity}
\title{Get reservoir capacity for concerned area}
\usage{
get_reservoir_capacity(area, opts = antaresRead::simOptions())
}
\arguments{
\item{area}{The area concerned by the simulation.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Get reservoir capacity for concerned area
}
