% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Reward.R
\name{get_Reward}
\alias{get_Reward}
\title{Get the reward matrix from simulations}
\usage{
get_Reward(
  simulation_res = NULL,
  simulation_names = NULL,
  pattern = NULL,
  district_name = "water values district",
  opts = antaresRead::simOptions(),
  correct_monotony = FALSE
)
}
\arguments{
\item{simulation_res}{list of the simulation names and constraint values to obtain the reward.}

\item{simulation_names}{generated by \code{runWaterValuesSimulation}}

\item{pattern}{A pattern to identify simulations.}

\item{district_name}{Name of the district used to store output.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{correct_monotony}{Binary argument (default to false). True to correct monotony of rewards.}
}
\value{
a data.table {timeid,MCyear,simulation overall cost}
}
\description{
Get the reward matrix from simulations
}
