% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWaterValuesSimulation.R
\name{getBackupData}
\alias{getBackupData}
\title{Generate backup data for area}
\usage{
getBackupData(area, mcyears, opts)
}
\arguments{
\item{area}{Character}

\item{mcyears}{Monte Carlo years to simulate or a vector of years indexes to launch.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
List with hydro_storage, load and misc_gen backups for area
}
\description{
Generate backup data for area
}
