% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterations_simulation_DP.R
\name{getCalculatedReward}
\alias{getCalculatedReward}
\title{Compare expected reward and Antares calculated reward}
\usage{
getCalculatedReward(reward, levels, expected_reward)
}
\arguments{
\item{reward}{Data frame containing estimation of the reward function,
same format as the output of \code{reward_offset}}

\item{levels}{Data frame with level (lev), optimal transition (opt) and transition
to evaluate (constraint) for each week (w)}

\item{expected_reward}{Data frame generated by \code{getExpectedReward}}
}
\value{
Data frame with expected reward, calculated reward and gap between them
}
\description{
Compare expected reward and Antares calculated reward
}
