% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterations_simulation_DP.R
\name{getExpectedReward}
\alias{getExpectedReward}
\title{Calculate the expected reward based on df_rewards for the new controls to evaluate,
used in \code{calculateBellmanWithIterativeSimulations}}
\usage{
getExpectedReward(reward, levels)
}
\arguments{
\item{reward}{Data frame containing estimation of the reward function,
same format as the output of \code{reward_offset}}

\item{levels}{Data frame with level (lev) and transition
to evaluate (constraint) for each week (w)}
}
\value{
Data frame {mcYear, week, u, reward}
}
\description{
Calculate the expected reward based on df_rewards for the new controls to evaluate,
used in \code{calculateBellmanWithIterativeSimulations}
}
