% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_trend.R
\name{getOptimalTrend}
\alias{getOptimalTrend}
\title{Calculate an optimal trajectory for the reservoir levels based on Bellman values and reward functions.}
\usage{
getOptimalTrend(
  level_init,
  watervalues,
  mcyears,
  reservoir_capacity,
  penalty_low,
  penalty_high,
  penalty_final_level,
  final_level,
  max_hydro_weekly,
  seed = 0,
  efficiency,
  mix_scenario = TRUE
)
}
\arguments{
\item{level_init}{Initial level of the reservoir in MWh generated by \code{get_initial_level_year_per_year()}.}

\item{watervalues}{Data frame aggregated watervalues generated by \code{Grid_Matrix()}.}

\item{mcyears}{Vector of integer. Monte Carlo years used to compute water values.}

\item{reservoir_capacity}{Double. Reservoir capacity for the given area in MWh given by \code{get_reservoir_capacity()}.}

\item{penalty_low}{Double. Penalty for violating the bottom rule curve, comparable to the unsupplied energy cost.}

\item{penalty_high}{Double. Penalty for violating the top rule curve, comparable to the spilled energy cost.}

\item{penalty_final_level}{Penalties (for both bottom and top rule curves) to force final level}

\item{final_level}{Double. Final level (in percent between 0 and 100) if final level is constrained. If you want initial level, use \code{get_initial_level()}.}

\item{max_hydro_weekly}{Maximum weekly pumping and generating power generated by the function \code{get_max_hydro()} with \code{timeStep="weekly"}.}

\item{seed}{If scenario are mixed, seed to make results reproducible.}

\item{efficiency}{Double between 0 and 1. Pumping efficiency ratio. Get it with \code{getPumpEfficiency()}.}

\item{mix_scenario}{Should scenario be mix from one week to another ?}
}
\value{
Data frame with level (\code{"lev"}) and optimal transition
(\code{"constraint"}) for each (\code{"week"}) and each (\code{"scenario"}).
For each (\code{"week"}), the (\code{"scenario"}) used is given by (\code{"mcYear"}).
}
\description{
Calculate an optimal trajectory for the reservoir levels based on Bellman values and reward functions.
}
