% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_storage_functions.R
\name{getPumpEfficiency}
\alias{getPumpEfficiency}
\title{Get pumping efficiency ratio}
\usage{
getPumpEfficiency(area, opts = antaresRead::simOptions())
}
\arguments{
\item{area}{Character. The Antares area concerned by water values computation.}

\item{opts}{List of study parameters returned by the function \code{antaresRead::setSimulationPath(simulation="input")} in input mode.}
}
\value{
Double. Pumping efficiency ratio.
}
\description{
Get pumping efficiency ratio for the given area. Reservoir management must be on.
}
