% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimulationNames.R
\name{getSimulationNames}
\alias{getSimulationNames}
\title{Utility function to get simulation's name}
\usage{
getSimulationNames(pattern, opts = antaresRead::simOptions())
}
\arguments{
\item{pattern}{A pattern to match among the simulation.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
A character vector.
}
\description{
Utility function to get simulation's name
}
\examples{
getSimulationNames("eco")
}
