% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBand.R
\name{getUnserverdEnergyCost}
\alias{getUnserverdEnergyCost}
\title{Get unserved energy cost for a given area}
\usage{
getUnserverdEnergyCost(area, opts = antaresRead::simOptions())
}
\arguments{
\item{area}{An 'antares' area.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
a numeric value
}
\description{
Get unserved energy cost for a given area
}
\examples{
\dontrun{

getUnserverdEnergyCost(area = "nl")

}
}
