% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeeklyMaxHydro.R
\name{getWeeklyMaxHydro}
\alias{getWeeklyMaxHydro}
\title{Get weekly maximum hydro storage energy}
\usage{
getWeeklyMaxHydro(area, opts = antaresRead::simOptions())
}
\arguments{
\item{area}{An 'antares' area.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
Maximum energy that can be generated by hydro storage over a week.
}
\description{
Get weekly maximum hydro storage energy
}
\examples{
\dontrun{

getWeeklyMaxHydro("fr")

}
}
