% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optim_H2_multi_nodes_functions.R
\name{grid_other_candidates}
\alias{grid_other_candidates}
\title{Compute the list of cluster candidates to study following their bounds}
\usage{
grid_other_candidates(candidates_data)
}
\arguments{
\item{candidates_data}{List of vector of doubles of length 3. One vector of double for each candidate.
The vectors of doubles have the form (bound min, bound max, number of points).}
}
\value{
a \code{list} of the capacity of each candidate cluster
}
\description{
Compute the list of cluster candidates to study following their bounds
}
