% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_storage_functions.R
\name{hydro_cost}
\alias{hydro_cost}
\title{Calculate  the hydro cost}
\usage{
hydro_cost(area, mcyears, simulation_name, opts)
}
\arguments{
\item{area}{A valid Antares area.}

\item{simulation_name}{simulation name in output folder.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
A real. the cost of hydro: stock Diff cost and hydro cost.
}
\description{
Calculate  the hydro cost
}
