% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{incr}
\alias{incr}
\title{Check if the vector has increasing values}
\usage{
incr(vector)
}
\arguments{
\item{vector}{numeric vector whose increasing check is wanted.}
}
\description{
Check if the vector has increasing values
}
