% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{interp_down}
\alias{interp_down}
\title{function used to replace the values under the level_low of reservoir}
\usage{
interp_down(maxi, statesid, q3)
}
\arguments{
\item{maxi}{the maximuum of water values}

\item{statesid}{the state Id of the the to replace}

\item{q3}{a quantile used in interpolation}
}
\description{
function used to replace the values under the level_low of reservoir
}
