% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWaterValuesSimulation.R
\name{launchSimulation}
\alias{launchSimulation}
\title{Launch the simulation}
\usage{
launchSimulation(
  opts,
  i,
  sim_name,
  path_solver,
  expansion,
  show_output_on_console,
  constraint_value
)
}
\arguments{
\item{opts}{List of study parameters returned by the function \code{antaresRead::setSimulationPath(simulation="input")} in input mode.}

\item{i}{Integer. Number of simulation.}

\item{sim_name}{Character. The name of the simulation.}

\item{path_solver}{Character containing the Antares Solver path, argument passed to \code{\link[antaresEditObject]{runSimulation}}.}

\item{expansion}{Binary. True if mode expansion (ie linear relaxation) of Antares is used to run simulations, argument passed to \code{\link[antaresEditObject]{runSimulation}}.
It is recommended to use mode expansion, it will be faster (only one iteration is done) and results will be smoother as the cost result will correspond to the linear relaxation of the problem.}

\item{show_output_on_console}{Argument passed to \code{\link[antaresEditObject]{runSimulation}}.}
}
\description{
Launch the simulation
}
\keyword{internal}
