% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optim_H2_multi_nodes_functions.R
\name{max_candidate}
\alias{max_candidate}
\title{Compute the candidate with maximal reward}
\usage{
max_candidate(grid_costs)
}
\arguments{
\item{grid_costs}{List. The firsts columns correspond to the candidates storage and cluster.
The last column contains the total reward under names "Total_cost".}
}
\value{
A \code{list} containing only the row of the best candidate from \code{grid_costs}.
}
\description{
Compute the candidate with maximal reward
}
