% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{mean_finite}
\alias{mean_finite}
\title{Calculate the mean of finite values.
Return \code{-Inf} if all \code{-Inf}.}
\usage{
mean_finite(x)
}
\arguments{
\item{x}{numeric vector whose mean is wanted.}
}
\description{
Calculate the mean of finite values.
Return \code{-Inf} if all \code{-Inf}.
}
\keyword{internal}
