% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanGridLayer.R
\name{meanGridLayer}
\alias{meanGridLayer}
\title{Calculate grid mean layer matrix}
\usage{
meanGridLayer(
  area,
  simulation_names,
  simulation_values = NULL,
  nb_cycle = 2L,
  district_name = "water values district",
  max_mcyears = NULL,
  week_53 = 0,
  method = c("mean-grid", "grid-mean"),
  states_steps = 0.05,
  reservoir_capacity = NULL,
  na_rm = FALSE,
  correct_outliers = FALSE,
  opts = antaresRead::simOptions()
)
}
\arguments{
\item{area}{An 'antares' area.}

\item{simulation_names}{Names of simulations to retrieve.}

\item{simulation_values}{Values for the simulation.}

\item{nb_cycle}{Number of times to run the algorithm.}

\item{district_name}{Name of the district used to store output.}

\item{max_mcyears}{Number of MC years to consider, by default all of them.}

\item{week_53}{Water values for week 53, by default 0.}

\item{method}{Perform mean grid algorithm or grid mean algorithm ?}

\item{states_steps}{Steps to discretize steps levels between the reservoir capacity and zero.}

\item{reservoir_capacity}{Reservoir capacity for the given area in GWh, if \code{NULL} (the default),
value in Antares is used if available else a prompt ask the user the value to be used.}

\item{na_rm}{Remove NAs}

\item{correct_outliers}{If TRUE, outliers in Bellman values are replaced by spline
interpolations. Defaults to FALSE.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
a \code{data.table}
}
\description{
Calculate grid mean layer matrix
}
\examples{
\dontrun{

opts <- antaresRead::setSimulationPath(path = "path/to/study", simulation = "input")

simulation_names <- getSimulationNames(pattern = "water_amount")
simulation_names <- gsub(pattern = ".*eco-", replacement = "", x = simulation_names)

simulation_values <- gsub("weekly_water_amount_", "", simulation_names)
simulation_values <- gsub(",", ".", simulation_values)
simulation_values <- as.numeric(simulation_values)

meanGridLayer(
  area = "fr",
  simulation_names = simulation_names,
  simulation_values = simulation_values,
  nb_cycle = 1,
  na_rm = TRUE,
  correct_outliers = TRUE
)

waterValuesViz(value_nodes_2017bis_1cycle)
waterValuesViz(value_nodes_2017bis_1cycle, add_band = TRUE, bandwidth = 100, failure_cost = 100)

}
}
