% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{monotonic_JM}
\alias{monotonic_JM}
\title{Force monotonic water values}
\usage{
monotonic_JM(results_dt)
}
\arguments{
\item{results_dt}{Output from \code{watervalues} or \code{Grid_Matrix}}
}
\value{
a \code{data.table}
}
\description{
Force monotonic water values
}
