% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{monotonic_JM2}
\alias{monotonic_JM2}
\title{Force monotonic water values Algorithm 2}
\usage{
monotonic_JM2(results_dt)
}
\arguments{
\item{results_dt}{Output from \code{watervalues} or \code{Grid_Matrix}}
}
\value{
a \code{data.table}
}
\description{
Force monotonic water values Algorithm 2
}
