% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{monotonic_VU}
\alias{monotonic_VU}
\title{Force monotonic water values}
\usage{
monotonic_VU(results_dt, noise_ratio = 1)
}
\arguments{
\item{results_dt}{Output from \code{watervalues} or \code{Grid_Matrix}}

\item{noise_ratio}{ratio to remove values that overcome the 95% quantile.}
}
\value{
a \code{data.table}
}
\description{
Force monotonic water values
}
