% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{names_reward}
\alias{names_reward}
\title{Extract values of turbined energy from simulation names.}
\usage{
names_reward(reward_dt, sim_name_pattern = "weekly_water_amount_")
}
\arguments{
\item{reward_dt}{reward data.table. Obtained using the function \code{get_Reward()}}

\item{sim_name_pattern}{the name of simulations used in \code{runWaterValuesSimulation()}}
}
\description{
Extract values of turbined energy from simulation names.
}
