% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optim_H2_multi_nodes_functions.R
\name{new_bounds}
\alias{new_bounds}
\title{Updates the candidates bounds at the end of an iteration.}
\usage{
new_bounds(best_candidate, candidates_data, storage_points)
}
\arguments{
\item{best_candidate}{_}

\item{candidates_data}{List of vector of doubles of length 3. One vector of double for each cluster candidate.
The vectors of doubles have the form (bound min, bound max, number of points).
The number of points must be >3 to update bounds at each iterations and approach solution.}

\item{storage_points}{Vector of integers. Storage volumes candidates at these iteration}
}
\value{
A \code{list} with : storage_bounds (Vector of integers. Storage volumes candidates at next iteration.) and candidates_data (List. It is of the same form as the \code{candidates_data} argument with new clusters candidates.)
}
\description{
Updates the candidates bounds at the end of an iteration.
}
