% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bellman_subfunctions.R
\name{next_week_value}
\alias{next_week_value}
\title{An intermediate function in bellman values calculation.
Select the next week bellman values of a position from all bellman values.}
\usage{
next_week_value(next_week_values_l, max_mcyear, i)
}
\arguments{
\item{next_week_values_l}{list of all bellman values of the next week.}

\item{max_mcyear}{the maximum number of scenarios.}

\item{i}{the number of row in the table of calculation.}
}
\description{
An intermediate function in bellman values calculation.
Select the next week bellman values of a position from all bellman values.
}
