% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{num_equal}
\alias{num_equal}
\title{Check if two numbers are equal.}
\usage{
num_equal(x, y, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{first number to compare.}

\item{y}{second number to compare.}

\item{tol}{the minimum difference from which two numbers are equal.}
}
\description{
Check if two numbers are equal.
}
