% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{plot_Bellman}
\alias{plot_Bellman}
\title{Plot Bellman and water values}
\usage{
plot_Bellman(value_nodes_dt, weeks_to_plot)
}
\arguments{
\item{value_nodes_dt}{A \code{dplyr::tibble()} containing the Bellman and water values. Output \code{aggregated_results} from \code{Grid_Matrix()}.}

\item{weeks_to_plot}{Vector of integer. Weeks to plot.}
}
\value{
A \code{ggplot2::ggplot()} object.
}
\description{
Plot Bellman values and water values for differents weeks listed in \code{weeks_to_plot}.
}
