% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{plot_flow}
\alias{plot_flow}
\title{Plot water flow Graph}
\usage{
plot_flow(
  area,
  timeStep = "weekly",
  mcyear = NULL,
  simulation_name = NULL,
  opts = antaresRead::simOptions(),
  shiny = F,
  ...
)
}
\arguments{
\item{area}{An 'antares' area.}

\item{timeStep}{Resolution of the data to import:
weekly (default, a linear interpolation is done on the data),
monthly (original data).}

\item{mcyear}{precise the MC year to plot.
all to plot all years.
Null plot the synthesis. Default NULL}

\item{simulation_name}{simulation name to plot.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
a \code{ggplot} object
}
\description{
Plot water flow Graph
}
