% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{plot_generation}
\alias{plot_generation}
\title{Plot Turbining power  Graph and return result table}
\usage{
plot_generation(
  area,
  timestep = "daily",
  Mcyear = NULL,
  min_path,
  max_path,
  simulation_name = NULL,
  opts = antaresRead::simOptions()
)
}
\arguments{
\item{area}{An 'antares' area.}

\item{min_path}{pathe of Pmin file "/user/Pmin **.txt"}

\item{simulation_name}{simulation name to plot.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{timeStep}{Resolution of the data to import:
weekly (default, a linear interpolation is done on the data),
monthly (original data).}

\item{mcyear}{precise the MC year to plot.
Null plot the synthesis. Default NULL}
}
\description{
Plot Turbining power  Graph and return result table
}
