% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{plot_reservoir}
\alias{plot_reservoir}
\title{Plot Reservoir Guide Graph and return result table}
\usage{
plot_reservoir(
  area,
  timeStep = "weekly",
  mcyear = NULL,
  simulation_name = NULL,
  opts = antaresRead::simOptions(),
  shiny = F
)
}
\arguments{
\item{area}{An 'antares' area.}

\item{timeStep}{Resolution of the data to import:
weekly (default, a linear interpolation is done on the data),
monthly (original data).}

\item{mcyear}{precise the MC year to plot.
all to plot all years.
Null plot the synthesis. Default NULL}

\item{simulation_name}{simulation name to plot.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{shiny}{Boolean. True to run the script in shiny mod.}
}
\value{
a \code{ggplot} object
}
\description{
Plot Reservoir Guide Graph and return result table
}
