% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{plot_results}
\alias{plot_results}
\title{Plot simulation variables comparison and real Ov. cost (for watervalues)}
\usage{
plot_results(
  simulations,
  district_name = "all",
  timeStep = "annual",
  mcyears,
  opts,
  plot_var,
  watervalues_areas,
  water_value = 50,
  ...
)
}
\arguments{
\item{simulations}{list of simulation names.}

\item{district_name}{district name that contains the all domain to study.}

\item{timeStep}{Resolution of the data to import.}

\item{mcyears}{precise the MC year to plot.
#' Null plot the synthesis. Default NULL}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{plot_var}{list of variables to plot.}

\item{watervalues_areas}{list of areas name that used water values.}
}
\value{
a \code{ggplot} object
}
\description{
Plot simulation variables comparison and real Ov. cost (for watervalues)
}
