% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{plot_reward}
\alias{plot_reward}
\title{Plot mean reward}
\usage{
plot_reward(reward_base, weeks_to_plot)
}
\arguments{
\item{reward_base}{A \code{dplyr::tibble()} containing reward functions. Output \code{reward} from \code{get_Reward()}.}

\item{weeks_to_plot}{Vector of integer. Weeks to plot.}
}
\value{
\item{graph}{A \code{ggplot2::ggplot()} object.}
\item{table}{A \code{dplyr::tibble()} containing plotted data.}
}
\description{
Plot mean reward functions for different weeks listed in \code{weeks_to_plot}.
}
