% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{plot_reward_mc}
\alias{plot_reward_mc}
\title{Plot reward}
\usage{
plot_reward_mc(reward_base, weeks_to_plot, scenarios_to_plot)
}
\arguments{
\item{reward_base}{A \code{dplyr::tibble()} containing reward functions. Output \code{reward} from \code{get_Reward()}.}

\item{weeks_to_plot}{Vector of integer. Weeks to plot.}

\item{scenarios_to_plot}{Vector of integer. Scenarios to plot.}
}
\value{
\item{graph}{A \code{ggplot2::ggplot()} object.}
\item{table}{A \code{dplyr::tibble()} containing plotted data.}
}
\description{
Plot reward functions for different weeks listed in \code{weeks_to_plot} and different scenarios listed in \code{scenarios_to_plot}.
}
