% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{post_process}
\alias{post_process}
\title{Fill the rest of reservoir states water values}
\usage{
post_process(
  results_dt,
  max_cost = 3000,
  min_cost = 0,
  full_imputation = FALSE,
  impute_method = "pmm",
  fix = F,
  min_vu = 0.5,
  max_vu = 1000,
  blocker = F
)
}
\arguments{
\item{results_dt}{Output from \code{watervalues} or \code{Grid_Matrix}.}

\item{max_cost}{maximal accepted water value (Replace by 'NA' crossed values)}

\item{min_cost}{minimal accepted water value (Replace by 'NA' crossed values)}

\item{full_imputation}{boolean. use 'max_cost' and 'min_cost' to impute the
missing values}

\item{impute_method}{impute method. check methods in \code{help(mice)}}

\item{fix}{Boolean. Fill missing values with constant values}

\item{min_vu}{minimal water value to use in fix strategy}

\item{max_vu}{maximal water value to use in fix strategy}

\item{blocker}{Boolean False to launch two times the post_process function recursively.}
}
\value{
a \code{data.table}
}
\description{
Fill the rest of reservoir states water values
}
