% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readReservoirLevels.R
\name{readReservoirLevelsV6}
\alias{readReservoirLevelsV6}
\title{Read Reservoir Levels for Antares version 6.}
\usage{
readReservoirLevelsV6(
  area,
  timeStep = "weekly",
  byReservoirCapacity = TRUE,
  opts = antaresRead::simOptions()
)
}
\arguments{
\item{area}{An 'antares' area.}

\item{timeStep}{Resolution of the data to import: weekly (default, a linear interpolation is done on the data), monthly (original data).}

\item{byReservoirCapacity}{Multiply the result by the reservoir capacity (if available).}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
a data.table
}
\description{
Read Reservoir Levels for Antares version 6.
}
\keyword{internal}
