% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optim_H2_multi_nodes_functions.R
\name{remove_candidate_cluster}
\alias{remove_candidate_cluster}
\title{Remove candidate cluster from the study before running investment process}
\usage{
remove_candidate_cluster(opts, candidates_types_gen)
}
\arguments{
\item{opts}{`list`
Study parameters (typically the output of `antaresRead::setSimulationPath(simulation = "input")`).}

\item{candidates_types_gen}{`data.frame`
A data.frame describing the generation candidates to be optimized.
**See section ‘Structure of candidates_types_gen’ for details about required columns and their usage.**}
}
\description{
Remove candidate cluster from the study before running investment process
}
