% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{remove_out}
\alias{remove_out}
\title{Post process water values}
\usage{
remove_out(
  results_dt,
  min = NULL,
  max = NULL,
  max_vu,
  min_vu,
  replace_na_method
)
}
\arguments{
\item{results_dt}{A \code{dplyr::tibble()} containing water values. Output \code{watervalues} from \code{Grid_Matrix()}.}

\item{min}{Double. Minimal accepted water value.}

\item{max}{Double. Maximal accepted water value.}

\item{max_vu}{Double. Value to use for water values higher than \code{max}.}

\item{min_vu}{Double. Value to use for water values lower than \code{min}.}

\item{replace_na_method}{Character. Method to replace extreme values, either \code{"constant values"} to replace
by \code{max_vu} and \code{min_vu} or \code{"extreme values"} to replace by the extreme values of the current week.}
}
\value{
A \code{dplyr::tibble()} with same format than \code{results_dt}.
}
\description{
Replace extreme water values by chosen values.
}
