% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{report_data}
\alias{report_data}
\title{Plot simulation variables comparison and real Ov. cost (for watervalues)}
\usage{
report_data(
  simulations,
  type = "area",
  district_list = "all",
  area_list = "all",
  timeStep = "annual",
  mcyears,
  opts,
  plot_var,
  watervalues_areas,
  ...
)
}
\arguments{
\item{simulations}{list of simulation names.}

\item{type}{"area" to import areas and "district" to import districts.}

\item{district_list}{list of district to plot. assign "all" to import all districts.}

\item{area_list}{list of area to plot. assign "all" to import all areas.
that contains the all domain to study.}

\item{timeStep}{Resolution of the data to import.}

\item{mcyears}{precise the MC year to plot.
#' Null plot the synthesis. Default NULL}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{plot_var}{list of variables to plot.}

\item{watervalues_areas}{list of areas name that used water values.}

\item{return_table}{boolean. return table or plot.}
}
\value{
a \code{ggplot} or \code{data.table} object
}
\description{
Plot simulation variables comparison and real Ov. cost (for watervalues)
}
