% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroStorage.R
\name{resetHydroStorage}
\alias{resetHydroStorage}
\title{Reset to 0 the hydro storage time series}
\usage{
resetHydroStorage(area, force = FALSE, opts = antaresRead::simOptions())
}
\arguments{
\item{area}{A valid Antares area.}

\item{force}{Force to reset the hydro storage file even if backup the file failed, see \code{Note}.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
An upddated list containing various information about the simulation.
}
\description{
Reset to 0 the hydro storage time series
}
\note{
The function make a copy of the original hydro storage time series,
 you can restore these with \code{restoreHydroStorage}.
}
\seealso{
\link{restoreHydroStorage}
}
