% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWaterValuesSimulation.R
\name{resetStudy}
\alias{resetStudy}
\title{Reset an Antares study. In case, there is a problem when executing \code{runWaterValuesSimulation},
run this function to restore the study.}
\usage{
resetStudy(
  opts,
  area,
  pumping,
  fictive_area = NULL,
  binding_constraint = "WeeklyWaterAmount"
)
}
\arguments{
\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{area}{The area concerned by the simulation}

\item{pumping}{Boolean. True to take into account the pumping.}

\item{fictive_area}{Name of the fictive area created}

\item{binding_constraint}{character}
}
\description{
Reset an Antares study. In case, there is a problem when executing \code{runWaterValuesSimulation},
run this function to restore the study.
}
