% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{restore_fictive_fatal_prod_demand}
\alias{restore_fictive_fatal_prod_demand}
\title{Restore load and misc gen time series}
\usage{
restore_fictive_fatal_prod_demand(
  area,
  opts = antaresRead::simOptions(),
  load,
  misc_gen
)
}
\arguments{
\item{area}{A valid Antares area.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{load}{Matrix with 8760 rows that contains backup load for the area}

\item{misc_gen}{Matrix with 8760 rows that contains backup misc generation for the area}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Restore load and misc gen time series
}
