% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_storage_functions.R
\name{restoreHydroStorage}
\alias{restoreHydroStorage}
\title{Restore the hydro storage time series}
\usage{
restoreHydroStorage(
  area,
  path = NULL,
  opts = antaresRead::simOptions(),
  silent = F
)
}
\arguments{
\item{area}{A valid Antares area.}

\item{path}{Path to a manual backup.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{silent}{Boolean. True to run without messages.}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Restore the hydro storage time series
}
