% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{restorePumpPower}
\alias{restorePumpPower}
\title{Restore the Pumping power series, used in \code{runWaterValuesSimulation}}
\usage{
restorePumpPower(
  area,
  path = NULL,
  opts = antaresRead::simOptions(),
  silent = F
)
}
\arguments{
\item{area}{A valid Antares area.}

\item{path}{Path to a manual backup.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{silent}{Boolean. True to run without messages.}
}
\value{
An updated list containing various information about the simulation
}
\description{
Restore the Pumping power series, used in \code{runWaterValuesSimulation}
}
