% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{restoreScenarioBuilder}
\alias{restoreScenarioBuilder}
\title{Restore initial scenario builder}
\usage{
restoreScenarioBuilder(opts, fictive_area)
}
\arguments{
\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{fictive_area}{Area for which to delete all scenario builder}
}
\description{
Restore initial scenario builder
}
