% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Reward.R
\name{reward_offset}
\alias{reward_offset}
\title{Modify local reward to take into account overall cost of the simulation}
\usage{
reward_offset(
  opts,
  df_reward,
  u0 = c(),
  mcyears,
  district_cost = "water values district"
)
}
\arguments{
\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{df_reward}{data.table computed by the function \code{get_local_reward}}

\item{u0}{Constraint values per week used in the simulation, empty list if none}

\item{mcyears}{Vector of years used to evaluate rewards}

\item{district_cost}{Name of district used to evaluate overall cost}
}
\value{
a data.table {mcYear,week,u,reward}
}
\description{
Modify local reward to take into account overall cost of the simulation
}
