% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWaterValuesSimulation.R
\name{runWaterValuesSimulation}
\alias{runWaterValuesSimulation}
\title{Run a simulation for calculating water values for a specific area}
\usage{
runWaterValuesSimulation(area, simulation_name = "WeeklyWaterAmount\%s",
  binding_constraint = "WeeklyWaterAmount", constraint_values,
  fictive_area = NULL, thermal_cluster = NULL, path_solver, wait = TRUE,
  show_output_on_console = TRUE, overwrite = FALSE,
  opts = antaresRead::simOptions())
}
\arguments{
\item{area}{The area concerned by the simulation.}

\item{simulation_name}{The name of the simulation, \code{s} is a placeholder for the constraint value,
this is usefull if \code{constraint_values} is a vector.}

\item{binding_constraint}{Name of the binding constraint.}

\item{constraint_values}{Vector of energy constraints on the link between the area and the fictive area.}

\item{fictive_area}{Name of the fictive area to create, argument passed to \code{\link{setupWaterValuesSimulation}}.}

\item{thermal_cluster}{Name of the thermal cluster to create, argument passed to \code{\link{setupWaterValuesSimulation}}.}

\item{path_solver}{Character containing the Antares Solver path, argument passed to \code{\link[antaresXpansion]{run_simulation}}.}

\item{wait}{Argument passed to \code{\link[antaresXpansion]{run_simulation}}.}

\item{show_output_on_console}{Argument passed to \code{\link[antaresXpansion]{run_simulation}}.}

\item{overwrite}{If area or cluster already exists, overwrite them ?}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Run a simulation for calculating water values for a specific area
}
