% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWaterValuesSimulation.R
\name{runWaterValuesSimulationMultiStock}
\alias{runWaterValuesSimulationMultiStock}
\title{Run Antares simulations in order to compute water values for multiple areas}
\usage{
runWaterValuesSimulationMultiStock(
  list_areas,
  list_pumping,
  list_efficiency,
  simulation_name = "wv_sim_\%s",
  mcyears = NULL,
  binding_constraint = "weekly_water_amount",
  path_solver = NULL,
  wait = TRUE,
  show_output_on_console = FALSE,
  overwrite = FALSE,
  opts = antaresRead::simOptions(),
  shiny = F,
  otp_dest = NULL,
  file_name = NULL,
  launch_simulations = T,
  reset_hydro = T,
  constraint_values = NULL,
  expansion = T,
  ...
)
}
\arguments{
\item{list_areas}{Vector of areas concerned by simulations.}

\item{list_pumping}{Named vector of binary to tell if pumping is available in areas.}

\item{list_efficiency}{Named vector of pumping efficiency.}

\item{simulation_name}{Character. The name of the simulation, \code{s} is a placeholder for the constraint value defined by \code{nb_disc_stock}.}

\item{mcyears}{Vector of integer. Monte Carlo years used to compute water values.}

\item{binding_constraint}{Character. Name of the binding constraint.}

\item{path_solver}{Character containing the Antares Solver path, argument passed to \code{\link[antaresEditObject]{runSimulation}}.}

\item{wait}{Argument passed to \code{\link[antaresEditObject]{runSimulation}}.}

\item{show_output_on_console}{Argument passed to \code{\link[antaresEditObject]{runSimulation}}.}

\item{overwrite}{If area or cluster already exists, should they be overwritten?}

\item{opts}{List of study parameters returned by the function \code{antaresRead::setSimulationPath(simulation="input")} in input mode.}

\item{shiny}{Boolean. True to run the script in shiny mod.}

\item{otp_dest}{the path in which the script save Rdata file.}

\item{file_name}{the Rdata file name.}

\item{launch_simulations}{Boolean. True to to run the simulations.}

\item{reset_hydro}{Boolean. True to reset hydro inflow to 0 before the simulation.}

\item{constraint_values}{Constraint values for each area to use to run simulations, generated by the function \code{\link{constraint_generator}} with an additional column \code{"area"}.}

\item{expansion}{Binary. True if mode expansion of Antares is used to run simulations, argument passed to \code{\link[antaresEditObject]{runSimulation}}.}
}
\value{
\item{simulation_names}{Vector of character. Names of simulations.}
\item{simulation_values}{A \code{dplyr::tibble()} generated by \code{constraint_generator()} describing controls for each simulation.}
\item{list_areas}{Parameter \code{list_areas}.}
\item{mc_years}{Parameter \code{mcyears}.}
\item{pumping}{Parameter \code{pumping}.}
\item{eff}{Parameter \code{efficiency}.}
\item{expansion}{Parameter \code{expansion}.}
\item{fictive_areas}{Parameter \code{fictive_areas}.}
}
\description{
Each simulation evaluates for each week and each scenario the cost of a set of controls for all areas in \code{list_areas}.
}
\seealso{
\code{runWaterValuesSimulation()}
}
