% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupWaterValuesSimulation.R
\name{setupWaterValuesSimulation}
\alias{setupWaterValuesSimulation}
\title{Setup a simulation before running it for calculating Water Values,
used in \code{runWaterValuesSimulation}}
\usage{
setupWaterValuesSimulation(
  area,
  fictive_area_name = paste0("watervalue_", area),
  thermal_cluster = "water_value_cluster",
  overwrite = FALSE,
  opts = antaresRead::simOptions(),
  pumping = F,
  backup
)
}
\arguments{
\item{area}{The area concerned by the simulation.}

\item{fictive_area_name}{Name of the fictive area to create.}

\item{thermal_cluster}{Name of the thermal cluster to create.}

\item{overwrite}{If area or cluster already exists, overwrite them ?}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{pumping}{Boolean. True to take into account the pumping.}

\item{backup}{List with hydro_storage, load and misc_gen backups for area
generated by the function \code{getBackupData}}
}
\value{
The result of antaresRead::simOptions
}
\description{
Setup a simulation before running it for calculating Water Values,
used in \code{runWaterValuesSimulation}
}
