% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R, R/test functions.R
\name{shiny_Grid_matrix}
\alias{shiny_Grid_matrix}
\title{Open watervalues Calculator in APP Web}
\usage{
shiny_Grid_matrix(simulation_res, opts = antaresRead::simOptions())

shiny_Grid_matrix(simulation_res, opts = antaresRead::simOptions())
}
\arguments{
\item{simulation_res}{List of simulation results returned by the function
\code{watervalues::runWaterValuesSimulation}}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Open watervalues Calculator in APP Web

Open watervalues Calculator in APP Web
}
