% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{shiny_water_values}
\alias{shiny_water_values}
\title{Open watervalues Calculator in APP Web}
\usage{
shiny_water_values(simulation_res = NULL, study_path, silent = F, ...)
}
\arguments{
\item{simulation_res}{List of simulation results returned by the function
\code{watervalues::runWaterValuesSimulation}}

\item{study_path}{the path of the Antares study}

\item{silent}{Boolean. TRUE to suppress warnings.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Open watervalues Calculator in APP Web
}
