% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_functions.R
\name{simulation_names_values}
\alias{simulation_names_values}
\title{simulations names and values generator}
\usage{
simulation_names_values(
  area,
  simulation_name,
  nb_disc_stock,
  pumping = F,
  opts = antaresRead::simOptions()
)
}
\arguments{
\item{area}{The area concerned by the simulation.}

\item{simulation_name}{The used name of the simulation.}

\item{nb_disc_stock}{Number of simulation launched.}

\item{pumping}{Boolean. True to take into account the pumping.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
simulations names and values generator
}
\note{
This function have no side effects on the Antares study used.
It's used in case you forget to save the output of script \code{runWaterValuesSimulation} to generate the output.
The results are meaningless unless you already launched the \code{runWaterValuesSimulation} with the same parameters.
}
