% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{states_to_percent}
\alias{states_to_percent}
\title{Convert Reservoir levels from MWh to percent of reservoir.}
\usage{
states_to_percent(data, states_step_ratio = 0.01)
}
\arguments{
\item{data}{A data.table contains the statesid and states columns}

\item{states_step_ratio}{percent step between two successive levels.}
}
\description{
Convert Reservoir levels from MWh to percent of reservoir.
}
\keyword{internal}
