% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{to_Antares_Format_bis}
\alias{to_Antares_Format_bis}
\title{Convert water values to Antares format with high accuracy}
\usage{
to_Antares_Format_bis(data)
}
\arguments{
\item{data}{A \code{dplyr::tibble()} representing water values, output \code{aggregated_results} generated by \code{Grid_Matrix()}.}
}
\value{
A 365*101 numeric matrix
}
\description{
This function converts water values generated by \code{Grid_Matrix()}
to the format expected by Antares: a 365*101 matrix, where
the rows are the 365 days of the year and the columns are round percentage values
ranging from 0 to 100 assessing the reservoir level.
Use \code{antaresEditObject::writeWaterValues()} to write water values.
}
\seealso{
\code{to_Antares_Format()}
}
