% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toAntaresFormat.R
\name{toAntaresFormat}
\alias{toAntaresFormat}
\title{Convert water values to Antares format}
\usage{
toAntaresFormat(data)
}
\arguments{
\item{data}{A 7-column data.table generated by \code{antaresWaterValues::meanGridLayer()}
or \code{antaresWaterValues::waterValues()}.}
}
\value{
A 365*101 numeric matrix
}
\description{
This function converts water values generated by \code{meanGridLayer} or
\code{waterValues} to the format expected by Antares: a 365*101 matrix, where
the rows are the 365 days of the year and the columns are round percentage values
ranging from 0 to 100 assessing the reservoir level.
Since \code{meanGridLayer} and \code{waterValues} output weekly values for an
arbitrary number of reservoir levels, interpolation is performed on both scales
in order to fit the desired format.
}
