% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optim_H2_multi_nodes_functions.R
\name{total_cost_loop}
\alias{total_cost_loop}
\title{Computes the total reward for a specific candidate following the reward function}
\usage{
total_cost_loop(
  area,
  mc_years,
  candidate_pool,
  candidates_types,
  storage_vol,
  df_reward,
  cvar,
  opts,
  penalty_low,
  penalty_high,
  penalty_final_level,
  storage_annual_cost,
  final_level,
  states_step_ratio
)
}
\arguments{
\item{area}{Character. Name of the area where the total reward is computed}

\item{mc_years}{Vector of integers. Monte Carlo years to compute reward}

\item{candidate_pool}{Vector of integers. Capacity of the cluster candidates}

\item{candidates_types}{Data_frame with column names : c("index", "name", "type", "TOTEX", "Marg_price").
It is a parameter of \code{MultiStock_H2_Investment_reward_compute_once}.}

\item{storage_vol}{Integer. Volume of the storage candidate}

\item{df_reward}{List. Rewards and decision space from the output of the function \code{antaresWaterValues::get_Reward()}.}

\item{cvar}{Numeric in [0,1]. The probability used in cvar algorithm.}

\item{opts}{List of study parameters returned by the function \code{antaresRead::setSimulationPath(simulation="input")} in input mode.}

\item{penalty_low}{Integer. Penalty for lower guide curve.}

\item{penalty_high}{Integer. Penalty for higher guide curve.}

\item{penalty_final_level}{Integer. Penalty for higher and lower final level.}

\item{storage_annual_cost}{Numeric. Total annual cost of the storage candidate in eur/MWh}

\item{final_level}{Numeric in [0, 100]. Final and initial level in percentage of H2 storage.}

\item{states_step_ratio}{Numeric. Step number for reward functions. Must be in adequacy with step ratio from \code{df_reward}.}
}
\value{
The total reward for the specified candidates in euros. Because of the Bellman values it is usually negative and large
}
\description{
Computes the total reward for a specific candidate following the reward function
}
