% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optim_H2_multi_nodes_functions.R
\name{total_cost_parallel_version}
\alias{total_cost_parallel_version}
\title{Computes the total reward for a list of candidate following the reward function, with parallel processing.
The storage volume is constant.
These function is called inside \code{parLapply()}.}
\usage{
total_cost_parallel_version(
  candidate_index,
  new_candidate_grid,
  node,
  mc_years_optim,
  candidates_types,
  storage_vol,
  list_rewards,
  nb_node,
  cvar,
  opts,
  penalty_low,
  penalty_high,
  penalty_final_level,
  storage_annual_cost,
  storage_final_level,
  step_number
)
}
\arguments{
\item{candidate_index}{Integer. Index of the candidate to study in \code{new_candidate_grid}. Argument passed to \code{parLapply()}}

\item{new_candidate_grid}{List of cluster candidates to study at these iteration.}

\item{node}{Character. Name of the area where the total reward is computed}

\item{mc_years_optim}{Vector of integers. Monte Carlo years to compute reward}

\item{candidates_types}{Data_frame with column names : c("index", "name", "type", "TOTEX", "Marg_price").
It is a parameter of \code{MultiStock_H2_Investment_reward_compute_once}.}

\item{storage_vol}{Integer. Volume of the storage candidate}

\item{list_rewards}{List. List of rewards and decision space from the output of the function \code{antaresWaterValues::get_Reward()}.
Rewards and decision spaces are given for each area marked by their number.}

\item{nb_node}{Integer. Number of the specific area to get corresponding reward from \code{list_rewards}}

\item{cvar}{Numeric in [0,1]. The probability used in cvar algorithm.}

\item{opts}{List of study parameters returned by the function \code{antaresRead::setSimulationPath(simulation="input")} in input mode.}

\item{penalty_low}{Integer. Penalty for lower guide curve.}

\item{penalty_high}{Integer. Penalty for higher guide curve.}

\item{penalty_final_level}{Integer. Penalty for higher and lower final level.}

\item{storage_annual_cost}{Numeric. Total annual cost of the storage candidate in eur/MWh}

\item{storage_final_level}{Numeric in [0, 100]. Final and initial level in percentage of H2 storage.}

\item{step_number}{Integer. Step number for reward functions.
Must be in adequacy with the step number if the reward function.}
}
\value{
The total reward for the specified candidates in euros. Because of the Bellman values it is usually negative and large
}
\description{
Computes the total reward for a list of candidate following the reward function, with parallel processing.
The storage volume is constant.
These function is called inside \code{parLapply()}.
}
