% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optim_H2_multi_nodes_functions.R
\name{update_reward_cluster_bande}
\alias{update_reward_cluster_bande}
\title{Modify the reward function to take into account the effects of a must-run cluster.}
\usage{
update_reward_cluster_bande(power, reward_init, mc_years)
}
\arguments{
\item{power}{Integer. Capacity of the cluster.}

\item{reward_init}{List. Rewards and decision space from the output of the function \code{antaresWaterValues::get_Reward()}.}

\item{mc_years}{Vector of integers. Monte Carlo years of the rewards.}
}
\value{
a \code{list} of rewards and decision space, like \code{reward_init}.
}
\description{
Modify the reward function to take into account the effects of a must-run cluster.
}
