% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optim_H2_multi_nodes_functions.R
\name{update_reward_cluster_flexible}
\alias{update_reward_cluster_flexible}
\title{Modify the reward function to take into account the effects of a non must-run cluster, with a speciifc capcaity and marginal cost.}
\usage{
update_reward_cluster_flexible(power, marg_cost, reward_init, mcYear)
}
\arguments{
\item{power}{Integer. Capacity of the cluster.}

\item{marg_cost}{Numeric. Marginal cost (and market bid) of the cluster in eur/MWh.}

\item{reward_init}{List. Rewards and decision space from the output of the function \code{antaresWaterValues::get_Reward()}.}

\item{mcYear}{Vector of integers. Monte Carlo years of the rewards.}
}
\value{
a \code{list} of rewards and decision space, like \code{reward_init}.
}
\description{
Modify the reward function to take into account the effects of a non must-run cluster, with a speciifc capcaity and marginal cost.
}
