% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_api.R
\name{validate_and_normalize_areas}
\alias{validate_and_normalize_areas}
\title{Validate a vector of area names against the study and normalize their case.}
\usage{
validate_and_normalize_areas(list_areas, opts)
}
\arguments{
\item{list_areas}{Character vector of area names.}

\item{opts}{List. Antares simulation options, as returned by \code{antaresRead::setSimulationPath()}.}
}
\value{
Character vector of \code{list_areas} normalized to lower case.
}
\description{
Validate a vector of area names against the study and normalize their case.
}
