% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizWVbyWeek.R
\name{vizWVbyWeek}
\alias{vizWVbyWeek}
\title{Visualize Water Weeks for a specified week}
\usage{
vizWVbyWeek(value_node, week, add_band = "none", ...)
}
\arguments{
\item{value_node}{Result from \code{meanGridLayer}.}

\item{week}{Week to visualize}

\item{add_band}{Add band around water values, choice between 'none', 'linear', 'spline' or 'both'.}

\item{...}{Parametters passed to \code{\link{addBand}}.}
}
\value{
A ggplot2 object
}
\description{
Visualize Water Weeks for a specified week
}
\examples{
\dontrun{

# TODO

}
}
