% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterValues.R
\name{waterValues}
\alias{waterValues}
\title{Run Water Values Simulation and Calculate Value Node}
\usage{
waterValues(area, nb_simulation = 10, n_runs = 3)
}
\arguments{
\item{area}{The area concerned by the simulation.}

\item{nb_simulation}{Number of simulation to launch, 
a vector of energy constraint will be created from 0 to 
the hydro storage maximum and of length this parameter.}

\item{n_runs}{Number of times to run the algorithm.}
}
\value{
a data.table
}
\description{
Run Water Values Simulation and Calculate Value Node
}
\examples{
\dontrun{

waterValues(area = "fr")

}
}
