% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterValuesViz.R
\name{waterValuesViz}
\alias{waterValuesViz}
\title{Visualize mean grid layer}
\usage{
waterValuesViz(
  value_nodes,
  add_band = FALSE,
  type = c("linear", "spline"),
  bandwidth = 100,
  failure_cost = 3000
)
}
\arguments{
\item{value_nodes}{Output from \code{waterValues} or \code{meanGridLayer}.}

\item{add_band}{Add band around water values, \code{TRUE} or \code{FALSE}.}

\item{type}{Perform a linear or a spline interpolation.}

\item{bandwidth}{Ith of the band to add.}

\item{failure_cost}{Cost of failure.}
}
\value{
a \code{ggplot} object
}
\description{
Visualize mean grid layer
}
