% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bellman.R
\name{Bellman}
\alias{Bellman}
\title{Compute Bellman values at step i from step i+1, used in \code{Grid_Matrix}}
\usage{
Bellman(
  Data_week,
  next_week_values_l,
  reward,
  E_max,
  P_max = 0,
  mcyears,
  cvar_value = 1,
  niveau_max,
  penalty_level_low,
  penalty_level_high,
  lvl_high,
  lvl_low,
  overflow_cost,
  next_state = NULL
)
}
\arguments{
\item{Data_week}{A "data.table" generated in Grid_Matrix code
that contains:
 * states Numeric. All the water values that can be set, listed in
  decreasing order.
 * value_inflow Numeric. Inflow values for each Monte-Carlo year.
 * Rewards for each simulation value and each Monte-Carlo year.
 * level_high Numeric. Highest possible reservoir value.
 * level_low Numeric. Lowest possible reservoir value.}

\item{next_week_values_l}{Numeric. Bellman values at step i+1.}

\item{reward}{Data.frame reward generated by \code{get_Reward()}}

\item{E_max}{Numeric of length 1. Maximum energy that can be generated by
hydro storage over one step of time.}

\item{P_max}{Numeric of length 1. Maximum energy that can be pumped to
reservoir over one step of time.}

\item{mcyears}{Vector. Monte Carlo years}

\item{cvar_value}{numeric in [0,1]. the probability used in cvar algorithm.}

\item{niveau_max}{Level max of the reservoir}

\item{penalty_level_low}{Penalty for violating the bottom rule curve, comparable to the unsupplied energy}

\item{penalty_level_high}{Penalty for violating the top rule curve, comparable to the spilled energy}

\item{lvl_high}{Double. Upper rule curve for the considered week.}

\item{lvl_low}{Double. Bottom rule curve for the considered week.}

\item{overflow_cost}{Cost for overflow (equal to spillage cost of the area)}

\item{next_state}{possible states for next week}
}
\value{
a \code{data.table} like Data_week with the Bellman values
}
\description{
Compute Bellman values at step i from step i+1, used in \code{Grid_Matrix}
}
\keyword{internal}
