% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bellman.R
\name{Bellman}
\alias{Bellman}
\title{Compute Bellman values at step i from step i+1, used in \code{Grid_Matrix}}
\usage{
Bellman(
  Data_week,
  next_week_values_l,
  decision_space,
  E_max,
  P_max = 0,
  method,
  mcyears,
  correct_outliers = FALSE,
  q_ratio = 0.75,
  counter,
  stop_rate = 5,
  debugger_feas = F,
  niveau_max,
  states_steps,
  penalty_level_low,
  penalty_level_high
)
}
\arguments{
\item{Data_week}{A "data.table" generated in Grid_Matrix code
that contains:
 * states Numeric. All the water values that can be set, listed in
  decreasing order.
 * value_inflow Numeric. Inflow values for each Monte-Carlo year.
 * Rewards for each simulation value and each Monte-Carlo year.
 * level_high Numeric. Highest possible reservoir value.
 * level_low Numeric. Lowest possible reservoir value.
 * states_next List of vectors enumerating all reachable states}

\item{next_week_values_l}{Numeric. Bellman values at step i+1.}

\item{decision_space}{Simulation constraints values}

\item{E_max}{Numeric of length 1. Maximum energy that can be generated by
hydro storage over one step of time.}

\item{P_max}{Numeric of length 1. Maximum energy that can be pumped to
reservoir over one step of time.}

\item{method}{Character. Perform mean of grids algorithm or grid of means algorithm or
grid of quantile algorithm.}

\item{mcyears}{Vector. Monte Carlo years}

\item{correct_outliers}{If TRUE, outliers in Bellman values are replaced by spline
interpolations. Defaults to FALSE.}

\item{q_ratio}{numeric in [0,1]. the probability used in quantile algorithm.}

\item{counter}{Numeric of length 1. number of the week in calculation.}

\item{stop_rate}{the percent from which the calculation stop. for example
\code{stop_rate=5} means the calculation stop if there is a week with less then
5\% accessibles states.}

\item{debugger_feas}{open debug mode in case there is an error of no accessible states}

\item{niveau_max}{Level max of the reservoir}

\item{states_steps}{Numeric. Discretization step of reservoir.}

\item{penalty_level_low}{Penalty for violating the bottom rule curve, comparable to the unsupplied energy}

\item{penalty_level_high}{Penalty for violating the top rule curve, comparable to the spilled energy}
}
\value{
a \code{data.table} like Data_week with the Bellman values
}
\description{
Compute Bellman values at step i from step i+1, used in \code{Grid_Matrix}
}
