% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optim_H2_multi_nodes_functions.R
\name{calculateRewardsSimulations}
\alias{calculateRewardsSimulations}
\title{Compute reward function with the 5 simulations method.
Called for each area of \code{MultiStock_H2_Investment_reward_compute_once}.
If they are several areas, the trajectories of the other storage are fixed on their optimal trend.}
\usage{
calculateRewardsSimulations(
  node,
  list_areas,
  list_efficiency,
  opts,
  mcyears,
  prefix,
  sim_number = 5,
  optimal_traj,
  candidates_types,
  list_max_hydro_weekly
)
}
\arguments{
\item{node}{Character. Name of the area where the reward is computed}

\item{list_areas}{Vector of characters of the names of areas to optimize.}

\item{list_efficiency}{Vector of numeric with pumping efficiency for the area with corresponding index in \code{list_areas}.}

\item{opts}{List of study parameters returned by the function \code{antaresRead::setSimulationPath(simulation="input")} in input mode.}

\item{mcyears}{Vector of integers. Monte Carlo years to run simulations}

\item{prefix}{Character. Prefix of the simulation.}

\item{sim_number}{Integer. Number of simulations.}

\item{optimal_traj}{Data frame containing optimal trajectory for all areas}

\item{candidates_types}{Data_frame with column names : c("index", "name", "type", "TOTEX", "Marg_price").}

\item{list_max_hydro_weekly}{List of data.frame. Generated by \code{get_max_hydro()} for each area.}
}
\value{
a \code{data_frame} containing the rewards returned by the function \code{antaresWaterValues::get_Reward()}
}
\description{
Compute reward function with the 5 simulations method.
Called for each area of \code{MultiStock_H2_Investment_reward_compute_once}.
If they are several areas, the trajectories of the other storage are fixed on their optimal trend.
}
