% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_storage_functions.R
\name{changeHydroManagement}
\alias{changeHydroManagement}
\title{Change hydro management}
\usage{
changeHydroManagement(watervalues = FALSE, heuristic = TRUE, opts, area)
}
\arguments{
\item{watervalues}{Binary. TRUE if use water values.}

\item{heuristic}{Binary. TRUE if use heuristic.}

\item{opts}{List of study parameters returned by the function \code{antaresRead::setSimulationPath(simulation="input")} in input mode.}

\item{area}{Character. Antares area for which to change hydro management.}
}
\description{
For the given area, choose if hydro storage is managed with water values and/or heuristic.
It is not possible to have neither water values nor heuristic.
}
