% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_storage_functions.R
\name{changeHydroManagement}
\alias{changeHydroManagement}
\title{Change the mode of management of an area}
\usage{
changeHydroManagement(watervalues = F, heuristic = T, opts, area)
}
\arguments{
\item{watervalues}{Binary. T if use watervalues}

\item{heuristic}{Binary. T if use heuristic}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{area}{Antares area}
}
\description{
Change the mode of management of an area
}
