% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_functions.R
\name{constraint_generator}
\alias{constraint_generator}
\title{Generate control values}
\usage{
constraint_generator(
  area,
  nb_disc_stock,
  pumping = TRUE,
  efficiency = NULL,
  opts,
  max_hydro = NULL,
  inflow = NULL,
  mcyears = NULL
)
}
\arguments{
\item{area}{Character. The Antares area concerned by water values computation.}

\item{nb_disc_stock}{Integer. Number of constraint values wanted for each week.}

\item{pumping}{Boolean. True to take into account the pumping capacity.}

\item{efficiency}{Double between 0 and 1. Pumping efficiency ratio. Get it with \code{getPumpEfficiency()}.}

\item{opts}{List of study parameters returned by the function \code{antaresRead::setSimulationPath(simulation="input")} in input mode.}

\item{max_hydro}{Weekly pumping and turbining maximum powers, generated by \code{get_max_hydro()}.}

\item{inflow}{Weekly inflow, generated by \code{get_inflow()}.}

\item{mcyears}{Vector of integer. Monte Carlo years used to compute water values.}
}
\value{
A \code{dplyr::tibble()} with 3 columns : \code{"week"} (integer), \code{"u"} (double), \code{"sim"} (character).
For each week from 1 to 52, \code{nb_disc_stock} different controls values are generated from maximum pumping power to maximum generating power.
The controls are equally spaced between maximum pumping power (negative value) and 0 and between 0 and maximum generating power (positive value).
Controls values are stored in column \code{"u"}. Column \code{"sim"} indexes control values in the form \code{"u_i"} with \code{i} the index.
}
\description{
Generate a list of control values, eg difference between final and initial level for each week for the given area.
Used to run simulations in \code{runWaterValuesSimulation()} and also to estimate reward functions
in functions such as \code{Grid_Matrix()} and \code{get_Reward()}.
}
