% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_functions.R
\name{constraint_generator}
\alias{constraint_generator}
\title{Generate the list of constraint values of the link between the fictive area and the real one}
\usage{
constraint_generator(
  area,
  nb_disc_stock,
  pumping = F,
  pumping_efficiency = NULL,
  opts,
  max_hydro = NULL,
  inflow = NULL
)
}
\arguments{
\item{area}{The area concerned by the simulation.}

\item{nb_disc_stock}{Number of simulation to launch, a vector of energy constraint.}

\item{pumping}{Boolean. True to take into account the pumping.}

\item{pumping_efficiency}{between 0 and 1. the pumping efficiency ratio.}

\item{opts}{List of simulation parameters returned by the function \code{antaresRead::setSimulationPath}}

\item{max_hydro}{weekly pumping and turbining maximum powers}

\item{inflow}{weekly inflow}
}
\description{
Generate the list of constraint values of the link between the fictive area and the real one
}
