% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{correct_concavity}
\alias{correct_concavity}
\title{Correct concavity of Bellman values to have nice monotony for watervalues}
\usage{
correct_concavity(results, weeks)
}
\arguments{
\item{results}{Intermediate results for Bellman values}

\item{weeks}{Weeks for which we want to correct concativity}
}
\value{
vector of corrected Bellman values
}
\description{
Correct concavity of Bellman values to have nice monotony for watervalues
}
