% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_reward}
\alias{example_reward}
\title{Example of output of \code{get_Reward}}
\format{
An object of class \code{Reward matrix , simulation names and values} of length 3.
}
\source{
Generated with a Antares toy study
}
\usage{
example_reward
}
\description{
A list containing a reward data frame, a list named simulation_names
and a data frame named simulation values
}
\keyword{datasets}
