% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_functions.R
\name{generate_constraints}
\alias{generate_constraints}
\title{This function generate binding constraints for \code{runWaterValuesSimulation}}
\usage{
generate_constraints(constraint_value, coeff, name_bc, opts)
}
\arguments{
\item{coeff}{the sens of the constraint notation in Antares.}

\item{name_bc}{the name of the constraint.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{constrain_value}{the value of the constraint}
}
\description{
This function generate binding constraints for \code{runWaterValuesSimulation}
}
