% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_functions.R
\name{generate_constraints}
\alias{generate_constraints}
\title{This function generate binding constraints for \code{runWaterValuesSimulation}}
\usage{
generate_constraints(
  constraint_value,
  coeff,
  name_constraint,
  efficiency = 0.75,
  opts,
  area = NULL
)
}
\arguments{
\item{constraint_value}{the value of the constraint}

\item{coeff}{the sens of the constraint notation in Antares.}

\item{name_constraint}{the name of the constraint.}

\item{efficiency}{in [0,1]. efficient ratio of pumping.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{area}{Area used to calculate watervalues}
}
\description{
This function generate binding constraints for \code{runWaterValuesSimulation}
}
