% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_functions.R
\name{generate_rhs_bc}
\alias{generate_rhs_bc}
\title{Modify time-series of clusters in fictive_area_bc to implement the constraint value
for each week and each MC year}
\usage{
generate_rhs_bc(constraint_value, coeff, opts)
}
\arguments{
\item{constraint_value}{Data.frame {week,sim,u}}

\item{coeff}{the sens and the name of constraints}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Modify time-series of clusters in fictive_area_bc to implement the constraint value
for each week and each MC year
}
