% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bellman_subfunctions.R
\name{get_bellman_values_interpolation}
\alias{get_bellman_values_interpolation}
\title{Create approximation of Bellman function for each scenario, used in \code{Bellman} and
in \code{getOptimalTrend}}
\usage{
get_bellman_values_interpolation(Data_week, next_week_values, mcyears)
}
\arguments{
\item{Data_week}{Data frame generated in \code{Grid_Matrix} code containing
list of states and years evaluated (we suppose there is only one week at a time)}

\item{next_week_values}{Vector of Bellman values to interpolate}

\item{mcyears}{Vector of monte carlo years to take into account}
}
\value{
List of \code{stats::approxfun} for each scenario
}
\description{
Create approximation of Bellman function for each scenario, used in \code{Bellman} and
in \code{getOptimalTrend}
}
