% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readReservoirLevels.R
\name{get_initial_level}
\alias{get_initial_level}
\title{Get initial level of an area}
\usage{
get_initial_level(area, opts)
}
\arguments{
\item{area}{An 'antares' area.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Get initial level of an area
}
