% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Reward.R
\name{get_local_reward}
\alias{get_local_reward}
\title{Compute reward functions based on one given simulation}
\usage{
get_local_reward(
  simu,
  possible_controls,
  max_hydro_hourly,
  area,
  mcyears,
  u0,
  district_balance,
  pump_eff
)
}
\arguments{
\item{simu}{List of simulation parameters returned by the function \code{antaresRead::setSimulationPath()} with the simulation selected.}

\item{possible_controls}{If \code{method_old=FALSE}, controls for which to compute reward, generated by \code{constraint_generator()}.}

\item{max_hydro_hourly}{Maximum hourly pumping and generating power generated by the function \code{get_max_hydro()} with \code{timeStep="hourly"}.}

\item{area}{Character. The Antares area concerned by water values computation.}

\item{mcyears}{Vector of integer. Monte Carlo years used to compute water values.}

\item{u0}{A \code{dplyr::tibble()} with columns \code{"week"}, \code{"u"} and \code{"mcYear"} (optional) that gives constraint values per week (and per scenario) used in the simulation.
Could be extracted from the \code{simulation_values} output of \code{runWaterValuesSimulation()} by filtering column \code{"sim"} to the corresponding simulation.}

\item{district_balance}{Character. Name of district used to evaluate controls on the stock.}
}
\value{
A \code{dplyr::tibble()} with columns \code{"week"}, \code{"mcYear"}, \code{"u"} and \code{"reward"}. Controls in column \code{"u"} correspond to \code{possible_controls}.
}
\description{
Reward functions are computed for all weeks from 1 to 52 and for all scenarios in \code{mcyears} for the given area.
For a specific week and a specific scenario, the reward function is evaluated for all controls listed in \code{possible_controls} based on marginal prices and on the real control done in the simulation.
This function is mainly used in \code{get_Reward()}.
}
