% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Reward.R
\name{get_local_reward}
\alias{get_local_reward}
\title{Calculate rewards for a simulation based on marginal prices, mainly used in \code{Get_Reward}}
\usage{
get_local_reward(
  opts,
  possible_controls,
  max_hydro,
  area_price,
  mcyears,
  u0,
  pump_eff = 1
)
}
\arguments{
\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{possible_controls}{data.frame {week,u} of controls evaluated per week}

\item{max_hydro}{data.frame {timeId,pump,turb} returned by the function \code{get_max_hydro}, should be hourly values}

\item{area_price}{Area used to evaluate marginal prices}

\item{mcyears}{Vector of years used to evaluate rewards}

\item{u0}{data.table {week,u} Constraint values per week used in the simulation}

\item{pump_eff}{Pumping efficiency}
}
\value{
a data.table {mcYear,week,u,reward}
}
\description{
Calculate rewards for a simulation based on marginal prices, mainly used in \code{Get_Reward}
}
