% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_max_hydro}
\alias{get_max_hydro}
\title{Get max hydro power that can be generated in a week}
\usage{
get_max_hydro(area, opts = antaresRead::simOptions())
}
\arguments{
\item{area}{The area concerned by the simulation.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Get max hydro power that can be generated in a week
}
