% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Reward.R
\name{get_Reward}
\alias{get_Reward}
\title{Get the reward matrix from simulations}
\usage{
get_Reward(
  simulation_names = NULL,
  ind = 1,
  pattern = NULL,
  district_name = "water values district",
  opts = antaresRead::simOptions()
)
}
\arguments{
\item{simulation_names}{Names of the simulation to obtain the reward.}

\item{ind}{index of the column of the 0 binding constraint simulation}

\item{pattern}{A pattern to identify simulations.}

\item{district_name}{Name of the district used to store output.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
a data.table {timeid,MCyear,simulation overall cost}
}
\description{
Get the reward matrix from simulations
}
