% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Reward.R
\name{get_Reward}
\alias{get_Reward}
\title{Get the reward matrix from simulations}
\usage{
get_Reward(
  simulation_values = NULL,
  simulation_names = NULL,
  pattern = NULL,
  district_name = "water values district",
  opts = antaresRead::simOptions(),
  correct_monotony = FALSE,
  method_old = TRUE,
  hours = 0:168,
  possible_controls = NULL,
  T_max,
  P_max,
  mcyears = "all",
  area = NULL,
  pump_eff = NULL,
  district_balance = "water values district",
  ind = NaN
)
}
\arguments{
\item{simulation_values}{Values generated by \code{runWaterValuesSimulation}}

\item{simulation_names}{generated by \code{runWaterValuesSimulation}}

\item{pattern}{A pattern to identify simulations.}

\item{district_name}{Name of the district used to store output.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{correct_monotony}{Binary argument (default to false). True to correct monotony of rewards.}

\item{method_old}{If T, linear interpolation used between simulations reward, else smarter interpolation based on marginal prices}

\item{hours}{If method_old=F, vector of hours used to evaluate costs/rewards of pumping/generating}

\item{possible_controls}{If method_old=F, vector of controls evaluated}

\item{T_max}{Max generating power}

\item{P_max}{Max pumping power}

\item{mcyears}{Vector of years used to evaluate rewards}

\item{area}{Area used to calculate watervalues}

\item{pump_eff}{Pumping efficiency}

\item{district_balance}{Name of district used to evaluate controls on the stock}

\item{ind}{name of simulation that corresponds to 0}
}
\value{
a data.table {timeid,MCyear,simulation overall cost}
}
\description{
Get the reward matrix from simulations
}
