% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Reward.R
\name{get_Reward}
\alias{get_Reward}
\title{Get the reward matrix from simulations, mainly used in \code{Grid_Matrix}}
\usage{
get_Reward(
  simulation_values = NULL,
  simulation_names = NULL,
  pattern = NULL,
  opts = antaresRead::simOptions(),
  correct_monotony = FALSE,
  method_old = TRUE,
  possible_controls = NULL,
  max_hydro,
  mcyears = "all",
  area = NULL,
  pump_eff = NULL,
  expansion = F
)
}
\arguments{
\item{simulation_values}{Values generated by \code{runWaterValuesSimulation}}

\item{simulation_names}{generated by \code{runWaterValuesSimulation}}

\item{pattern}{A pattern to identify simulations.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{correct_monotony}{Binary argument (default to false). True to correct monotony of rewards.}

\item{method_old}{If T, linear interpolation used between simulations reward, else smarter interpolation based on marginal prices}

\item{possible_controls}{If method_old=F, data.frame {week,u} of controls evaluated per week}

\item{max_hydro}{data.frame {timeId,pump,turb} returned by the function \code{get_max_hydro}, should be hourly values}

\item{mcyears}{Vector of years used to evaluate rewards}

\item{area}{Area used to calculate watervalues}

\item{pump_eff}{Pumping efficiency}

\item{expansion}{Binary. True if mode expansion was used to run simulations}
}
\value{
list containing a data.table {timeid,MCyear,simulation overall cost}, list of simulations names and list of simulations values
}
\description{
Get the reward matrix from simulations, mainly used in \code{Grid_Matrix}
}
