% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_weekly_cost}
\alias{get_weekly_cost}
\title{Get overall cost in a week, used in different functions}
\usage{
get_weekly_cost(opts = antaresRead::simOptions(), mcyears, expansion = F)
}
\arguments{
\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{mcyears}{Vector of years used to evaluate cost}

\item{expansion}{Binary. True if mode expansion was used to run simulations}
}
\description{
Get overall cost in a week, used in different functions
}
