% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWaterValuesSimulation.R
\name{getBackupData}
\alias{getBackupData}
\title{Generate backup data for area}
\usage{
getBackupData(area, mcyears, opts)
}
\arguments{
\item{area}{Character. The Antares area concerned by water values computation.}

\item{mcyears}{Vector of integer. Monte Carlo years used to compute water values.}

\item{opts}{List of study parameters returned by the function \code{antaresRead::setSimulationPath(simulation="input")} in input mode.}
}
\value{
List with hydro_storage, load and misc_gen backups for area
}
\description{
Generate backup data for area
}
\keyword{internal}
