% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistock_sequential_plaia.R
\name{getBellmanValuesSequentialMultiStockWithPlaia}
\alias{getBellmanValuesSequentialMultiStockWithPlaia}
\title{Calculate Bellman values for several areas area per area.
For one area, reward functions are computed with \code{calculateRewardsSimulationsWithPlaia()} and
Bellman values are computed with \code{Grid_Matrix()}. Finally, an optimal trajectory is computed for
the area with \code{getOptimalTrend}. The trajectory is used when computing reward functions for the next areas.
Initially, when Bellman values have been computed for an area, either \code{initial_traj} is used or a short-term
trajectory is used.}
\usage{
getBellmanValuesSequentialMultiStockWithPlaia(
  list_areas,
  opts,
  nb_simulations,
  mcyears,
  penalty_low,
  penalty_high,
  states_step_ratio = 1/50,
  cvar_value = 1,
  penalty_final_level = NULL,
  initial_traj = NULL,
  list_areas_to_compute = NULL
)
}
\arguments{
\item{list_areas}{Vector of areas concerned by simulations.}

\item{opts}{List of study parameters returned by the function \code{antaresRead::setSimulationPath(simulation="input")} in input mode.}

\item{nb_simulations}{Number of controls to simulate}

\item{mcyears}{Vector of integer. Monte Carlo years used to compute water values.}

\item{penalty_low}{Double. Penalty for violating the bottom rule curve, comparable to the unsupplied energy cost.}

\item{penalty_high}{Double. Penalty for violating the top rule curve, comparable to the spilled energy cost.}

\item{states_step_ratio}{Double. Discretization ratio to generate steps levels
between the reservoir capacity and zero for which Bellman values are computed.}

\item{cvar_value}{Double from 0 to 1. The probability used in cvar method.}

\item{penalty_final_level}{Penalties (for both bottom and top rule curves) to force final level}

\item{initial_traj}{Initial trajectory (used for other storages)}

\item{list_areas_to_compute}{Vector of character. Areas for which to compute Bellman values. If \code{NULL}, all areas in \code{list_areas} are used.}
}
\value{
List containing aggregated water values, reward functions and optimal trajectories.
}
\description{
Calculate Bellman values for several areas area per area.
For one area, reward functions are computed with \code{calculateRewardsSimulationsWithPlaia()} and
Bellman values are computed with \code{Grid_Matrix()}. Finally, an optimal trajectory is computed for
the area with \code{getOptimalTrend}. The trajectory is used when computing reward functions for the next areas.
Initially, when Bellman values have been computed for an area, either \code{initial_traj} is used or a short-term
trajectory is used.
}
