% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watervalues_plaia.R
\name{getBellmanValuesWithPlaia}
\alias{getBellmanValuesWithPlaia}
\title{Compute Bellman values for several storage with plaia}
\usage{
getBellmanValuesWithPlaia(
  opts,
  mcyears,
  list_areas,
  force_final_level,
  penalty_final_level,
  penalty_low,
  penalty_high,
  name_sim,
  n_controls
)
}
\arguments{
\item{opts}{List of study parameters returned by the function \code{antaresRead::setSimulationPath(simulation="input")} in input mode.}

\item{mcyears}{Vector of integer. Monte Carlo years used to compute water values.}

\item{list_areas}{Vector of areas concerned by simulations.}

\item{force_final_level}{Binary. Whether final level should be constrained.}

\item{penalty_final_level}{Double. Penalties for both rule curves to constrain final level.}

\item{penalty_low}{Double. Penalty for violating the bottom rule curve, comparable to the unsupplied energy cost.}

\item{penalty_high}{Double. Penalty for violating the top rule curve, comparable to the spilled energy cost.}

\item{name_sim}{Character. Name of the simulation to launch}

\item{n_controls}{Integer. Number of controls to evaluate reward functions.}
}
\value{
List of watervalues matrix
}
\description{
For each storage, reward functions are evaluated and Bellman values are computed with plaia.
Results are written directly to the output folder.
}
