% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterations_simulation_DP.R
\name{getInitialTrend}
\alias{getInitialTrend}
\title{Calculate a trajectory for the reservoir levels inside rule curves
taking into account the mean inflow, used in \code{calculateBellmanWithIterativeSimulations}}
\usage{
getInitialTrend(
  level_init,
  controls,
  inflow,
  mcyears,
  niveau_max,
  penalty_low,
  penalty_high,
  reservoir
)
}
\arguments{
\item{level_init}{Initial level of the reservoir in MWh}

\item{controls}{Data frame containing possible transition for each week,
generated by the function \code{constraint_generator}}

\item{inflow}{Data frame with inflows for each week and each scenario,
generated by the function \code{antaresRead::readInputTS}}

\item{mcyears}{MC years to take into account}

\item{niveau_max}{Capacity of the reservoir in MWh}

\item{penalty_low}{Penalty for violating the bottom rule curve}

\item{penalty_high}{Penalty for violating the top rule curve}

\item{reservoir}{Data frame with weekly rule curves with values between 0 and 1,
generated by the function \code{readReservoirLevels}}
}
\value{
Data frame with level (lev), optimal transition (opt) and transition
to evaluate (constraint) for each week (w)
}
\description{
Calculate a trajectory for the reservoir levels inside rule curves
taking into account the mean inflow, used in \code{calculateBellmanWithIterativeSimulations}
}
