% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterations_simulation_DP.R
\name{getOptimalTrend}
\alias{getOptimalTrend}
\title{Calculate an optimal trajectory for the reservoir levels based on water values
taking into account the mean inflow,
used in \code{calculateBellmanWithIterativeSimulations}}
\usage{
getOptimalTrend(
  level_init,
  watervalues,
  mcyears,
  reward,
  controls,
  niveau_max,
  df_levels,
  penalty_low,
  penalty_high,
  method_fast = F,
  max_hydro_weekly,
  n = 0
)
}
\arguments{
\item{level_init}{Initial level of the reservoir in MWh}

\item{watervalues}{Data frame aggregated watervalues generated by \code{Grid_Matrix}}

\item{mcyears}{Vector of monte carlo years used to evaluate rewards}

\item{reward}{Data frame containing estimation of the reward function,
same format as the output of \code{reward_offset}}

\item{controls}{Data frame containing possible transition for each week,
generated by the function \code{constraint_generator}}

\item{niveau_max}{Capacity of the reservoir in MWh}

\item{df_levels}{Data frame containing all previous evaluated controls, same format
as \code{getOptimalTrend} and \code{getInitialTrend}}

\item{penalty_low}{Penalty for violating the bottom rule curve}

\item{penalty_high}{Penalty for violating the top rule curve}

\item{method_fast}{Method to choose evaluated controls}

\item{max_hydro_weekly}{Data frame with weekly maximum pumping and generating powers}

\item{n}{Iteration}
}
\value{
Data frame with level (lev) and transition
to evaluate (constraint) for each week (w)
}
\description{
Calculate an optimal trajectory for the reservoir levels based on water values
taking into account the mean inflow,
used in \code{calculateBellmanWithIterativeSimulations}
}
