% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterations_simulation_DP.R
\name{getOptimalTrend}
\alias{getOptimalTrend}
\title{Calculate an optimal trajectory for the reservoir levels based on water values
taking into account the mean inflow,
used in \code{calculateBellmanWithIterativeSimulations}}
\usage{
getOptimalTrend(
  level_init,
  watervalues,
  mcyears,
  reward,
  controls,
  niveau_max,
  df_levels,
  penalty_low,
  penalty_high,
  penalty_final_level,
  final_level,
  max_hydro_weekly,
  n = 0,
  pump_eff,
  mix_scenario = TRUE,
  df_previous_cut = NULL
)
}
\arguments{
\item{level_init}{Initial level of the reservoir in MWh}

\item{watervalues}{Data frame aggregated watervalues generated by \code{Grid_Matrix}}

\item{mcyears}{Vector of monte carlo years used to evaluate rewards}

\item{reward}{Data frame containing estimation of the reward function,
same format as the output of \code{reward_offset}}

\item{controls}{Data frame containing possible transition for each week,
generated by the function \code{constraint_generator}}

\item{niveau_max}{Capacity of the reservoir in MWh}

\item{df_levels}{Data frame containing all previous evaluated controls, same format
as \code{getOptimalTrend}}

\item{penalty_low}{Penalty for violating the bottom rule curve}

\item{penalty_high}{Penalty for violating the top rule curve}

\item{penalty_final_level}{Penalty for final level}

\item{final_level}{Final level}

\item{max_hydro_weekly}{Data frame with weekly maximum pumping and generating powers}

\item{n}{Iteration}

\item{pump_eff}{Pumping efficiency (1 if no pumping)}

\item{mix_scenario}{Should scenario be mix from one week to another ?}

\item{df_previous_cut}{Data frame containing previous estimations of cuts}
}
\value{
Data frame with level (lev) and transition
to evaluate (constraint) for each week (w)
}
\description{
Calculate an optimal trajectory for the reservoir levels based on water values
taking into account the mean inflow,
used in \code{calculateBellmanWithIterativeSimulations}
}
