% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid_Matrix.R
\name{Grid_Matrix}
\alias{Grid_Matrix}
\title{Calculate grid layer matrix of Bellman values and water values}
\usage{
Grid_Matrix(
  area,
  simulation_names,
  simulation_values = NULL,
  nb_cycle = 1L,
  district_name = "water values district",
  max_mcyears = NULL,
  week_53 = 0,
  states_step_ratio = 0.01,
  reservoir_capacity = NULL,
  na_rm = FALSE,
  overwrite = TRUE,
  correct_outliers = FALSE,
  method,
  only_input = FALSE,
  q_ratio = 0.5,
  monotonic_bellman = FALSE,
  test_week = NULL,
  opts = antaresRead::simOptions(),
  shiny = F,
  ...
)
}
\arguments{
\item{area}{An 'antares' area.}

\item{simulation_names}{Names of simulations to retrieve.}

\item{simulation_values}{Values for the simulation.}

\item{nb_cycle}{Number of times to run the algorithm.}

\item{district_name}{Name of the district used to store output.}

\item{max_mcyears}{Number of MC years to consider, by default all of them.}

\item{week_53}{Water values for week 53, by default 0.}

\item{states_step_ratio}{Discretization ratio to generate steps levels
between the reservoir capacity and zero . Defaults to 0.05}

\item{reservoir_capacity}{Reservoir capacity for the given area in GWh, if \code{NULL} (the default),
value in Antares is used if available else a prompt ask the user the value to be used.}

\item{na_rm}{Remove NAs}

\item{correct_outliers}{If TRUE, outliers in Bellman values are replaced by spline
interpolations. Defaults to FALSE.}

\item{method}{Perform mean grid algorithm or grid mean algorithm ?}

\item{only_input}{if TRUE skip bellman values calculation and return the input
provided to calculate. Used mainly to verify for tests. Default FALSE}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
a \code{data.table}
}
\description{
Calculate grid layer matrix of Bellman values and water values
}
