% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid_Matrix.R
\name{Grid_Matrix}
\alias{Grid_Matrix}
\title{Calculate grid layer matrix of Bellman values and water values}
\usage{
Grid_Matrix(
  area,
  simulation_names,
  simulation_values = NULL,
  nb_cycle = 1L,
  district_name = "water values district",
  mcyears = NULL,
  week_53 = 0,
  states_step_ratio = 0.01,
  reservoir_capacity = NULL,
  na_rm = FALSE,
  correct_outliers = FALSE,
  method,
  only_input = FALSE,
  q_ratio = 0.5,
  monotonic_bellman = FALSE,
  test_week = NULL,
  parallel = FALSE,
  opts = antaresRead::simOptions(),
  shiny = F,
  inaccessible_states = F,
  until_convergence = F,
  convergence_rate = 0.9,
  convergence_criteria = 1,
  cycle_limit = 10,
  ...
)
}
\arguments{
\item{area}{An 'antares' area.}

\item{simulation_names}{Names of simulations to retrieve.}

\item{simulation_values}{Values for the simulation.}

\item{nb_cycle}{Number of times to run the algorithm.}

\item{district_name}{Name of the district used to store output.}

\item{mcyears}{MC years to consider, by default all of them.}

\item{week_53}{Water values for week 53, by default 0.}

\item{states_step_ratio}{Discretization ratio to generate steps levels
between the reservoir capacity and zero . Defaults to 0.05}

\item{reservoir_capacity}{Reservoir capacity for the given area in MWh,
if \code{NULL} (the default), value in Antares is used if available else
a prompt ask the user the value to be used.}

\item{na_rm}{Remove NAs}

\item{correct_outliers}{If TRUE, outliers in Bellman values are replaced
by spline interpolations. Defaults to FALSE.}

\item{method}{Perform mean of grids algorithm or grid of means algorithm or
grid of quantile algorithm.}

\item{only_input}{if TRUE skip bellman values calculation and return the input}

\item{q_ratio}{from 0 to 1. the probability used in quantile method
to determine a bellman value which q_ratio all bellman values are equal or
less to it. (quantile(q_ratio))}

\item{monotonic_bellman}{force increasing bellman values with the stock
level in the calculation.}

\item{test_week}{the week number u want to see it's calculation information
in the console}

\item{parallel}{Boolean. True to use parallel computing.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{shiny}{Boolean. True to run the script in shiny mod.}

\item{inaccessible_states}{Boolean. True to delete inaccessible states of
any scenario in the result.}

\item{until_convergence}{Boolean.TRUE to repeat cycle until convergence or
attending the limit.}

\item{convergence_rate}{from 0 to 1. Define the convergence level from which
we suppose that no need to continue another cycle..}

\item{convergence_criteria}{the value define convergence. if the difference
between two water values is less then this value those values are converged.}

\item{cycle_limit}{Define cycles limit when you are in the until_convergence mod.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a \code{data.table}
}
\description{
Calculate grid layer matrix of Bellman values and water values
}
