% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWaterValuesSimulation.R
\name{launchSimulation}
\alias{launchSimulation}
\title{Launch the simulation}
\usage{
launchSimulation(
  opts,
  i,
  sim_name,
  path_solver,
  expansion,
  show_output_on_console,
  constraint_value
)
}
\arguments{
\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{i}{Integer. Number of simulation.}

\item{sim_name}{Character. The name of the simulation.}

\item{path_solver}{Character containing the Antares Solver path, argument passed to \code{\link[antaresEditObject]{runSimulation}}.}

\item{expansion}{Binary. True if mode expansion was used to run simulations}

\item{show_output_on_console}{Argument passed to \code{\link[antaresEditObject]{runSimulation}}.}

\item{constraint_value}{Data.frame {week,sim,u}}
}
\description{
Launch the simulation
}
