% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{mean_or_inf}
\alias{mean_or_inf}
\title{Calculate the mean if there is no infinite or missing value.
Return \code{-Inf} in the other case.}
\usage{
mean_or_inf(x)
}
\arguments{
\item{x}{numeric vector whose mean is wanted.}
}
\description{
Calculate the mean if there is no infinite or missing value.
Return \code{-Inf} in the other case.
}
