% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanGridLayer.R
\name{meanGridLayer}
\alias{meanGridLayer}
\title{Calculate grid mean layer matrix}
\usage{
meanGridLayer(area, simulation_names, simulation_values = NULL, n_runs = 2L,
  district_name = "water values district", max_mcyears = NULL,
  week_53 = 0, method = c("mean-grid", "grid-mean"), states_steps = 0.05,
  reservoir_capacity = NULL, na_rm = FALSE,
  opts = antaresRead::simOptions())
}
\arguments{
\item{area}{An 'antares' area.}

\item{simulation_names}{Names of simulations to retrieve.}

\item{simulation_values}{Values for the simulation.}

\item{n_runs}{Number of times to run the algoithm.}

\item{district_name}{Name of the district used to store output.}

\item{max_mcyears}{Number of MC years to consider, by default all of them.}

\item{week_53}{Water values for week 53, by default 0.}

\item{method}{Perform mean grid algorithm or grid mean algorithm ?}

\item{states_steps}{Steps to discretize steps levels between the reservoir capacity and zero.}

\item{reservoir_capacity}{Reservoir capacity for the given area, if \code{NULL} (the default),
value in Antares is used.}

\item{na_rm}{Remove NAs}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
a \code{data.table}
}
\description{
Calculate grid mean layer matrix
}
\examples{
\dontrun{

# TODO

}
}
