% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optim_H2_multi_nodes_functions.R
\name{MultiStock_H2_Investment_reward_compute_once}
\alias{MultiStock_H2_Investment_reward_compute_once}
\title{Compute optimal candidates for H2 system}
\usage{
MultiStock_H2_Investment_reward_compute_once(
  areas_invest,
  max_ite,
  storage_bounds,
  storage_points_nb,
  candidates_types_gen,
  penalty_low = 5000,
  penalty_high = 5000,
  penalty_final_level = 5000,
  opts,
  mc_years_optim,
  path_to_antares,
  study_path,
  cvar = 1,
  storage_annual_cost,
  launch_sims = T,
  nb_sims = 51,
  parallelprocess = F,
  nb_sockets = 0,
  unspil_cost = 3000,
  edit_study = F,
  back_to_first_node = F
)
}
\arguments{
\item{areas_invest}{Vector of characters of the names of areas to optimize.}

\item{max_ite}{Integer. Maximum number of iterations for each area.}

\item{storage_bounds}{Vector of integers of length 2, with the form (min, max).}

\item{storage_points_nb}{Integer. Number of storage points to test at each iteration.
Must be >3 to update bounds at each iterations and approach solution.}

\item{candidates_types_gen}{Data_frame with column names : \code{c(index, name, type, TOTEX, Marg_price, Part_fixe, Prix_fixe,
Borne_min, Borne_max, Points_nb, Zone)}. Each row describes a cluster candidate. The index should correspond to the index of the candidate in \code{candidates_data}.
The name is a character, the type is either \code{"cluster_flexible"} or \code{"cluster_bande"}, TOTEX is in eur/MW/year, Marg_price is in eur/MWh.
Part_fixe is the fixed part for a variable cluster between 0 and 1, Prix_fixe in eur/MWh is its price.
Borne_min and Borne_max are in MW, Points_nb is an integer (number of candidates tested at each iteration, it should be at least 4).
Zone is a character containing the name of the area where to propose the candidate.}

\item{penalty_low}{Integer. Penalty for lower guide curve.}

\item{penalty_high}{Integer. Penalty for higher guide curve.}

\item{penalty_final_level}{Integer. Penalty for higher and lower final level.}

\item{opts}{List of study parameters returned by the function \code{antaresRead::setSimulationPath(simulation="input")} in input mode.}

\item{mc_years_optim}{Vector of integers. Monte Carlo years to perform the optimization.}

\item{path_to_antares}{Character containing the Antares Solver path, argument passed to \code{\link[antaresEditObject]{runSimulation}}.}

\item{study_path}{Character. Path to the simulation, argument passed to \code{antaresRead::setSimulationPath}.}

\item{cvar}{Numeric in [0,1]. The probability used in cvar algorithm.}

\item{storage_annual_cost}{Numeric. Annual cost of storage in eur/MWh.}

\item{launch_sims}{Boolean. True to launch simulations at each iterations. False if simulations already run.}

\item{nb_sims}{Integer. Number of simulations to launch to evaluate reward.}

\item{parallelprocess}{Boolean. True to compute Water values with parallel processing.}

\item{nb_sockets}{Integer. Number of sockets for parallel processing}

\item{unspil_cost}{Numeric. Unspilled energy cost in eur/MW for all concerned areas.}

\item{edit_study}{Boolean. True to edit study with optimal candidates.}

\item{back_to_first_node}{Boolean. True to play again first node at the end. There is no possibility to go uninvest.}
}
\value{
a \code{list} containing for each area detailed results (best candidate, all total costs, reward function, optimization time)
}
\description{
Search optimal solution for bounded candidates for the H2 systems including storage, must-run clusters and flexibles clusters.
}
