% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{plot_Bellman}
\alias{plot_Bellman}
\title{Plot Bellman and Water values}
\usage{
plot_Bellman(
  value_nodes_dt,
  week_number,
  penalty_low = 10000,
  penalty_high = 10000,
  force_final_level = F,
  penalty_final_level = 0
)
}
\arguments{
\item{value_nodes_dt}{A data.table contains the Bellman and water values .
Obtained using the function Grid_Matrix()}

\item{week_number}{Numeric of length 1. number of the week to plot.}

\item{penalty_low}{Penalty for the lower rule curve}

\item{penalty_high}{Penalty for the higher rule curve}

\item{force_final_level}{Binary. Whether final level should be constrained}

\item{penalty_final_level}{Penalties (for both bottom and top rule curves) to constrain final level}
}
\value{
a \code{ggplot} object
}
\description{
Plot Bellman and Water values
}
