% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{plot_Bellman}
\alias{plot_Bellman}
\title{Plot Bellman and Water values}
\usage{
plot_Bellman(
  value_nodes_dt,
  week_number,
  param = "vu",
  states_step_ratio = 0.01,
  bellman_week = NULL,
  ...
)
}
\arguments{
\item{value_nodes_dt}{A data.table contains the Bellman and water values .
Obtained using the function Grid_Matrix()}

\item{week_number}{Numeric of length 1. number of the week to plot.}

\item{param}{string contains the element to plot
* "vu" to plot only water values
* "b" to plot only bellman values
* "both" to plot both water and bellman values
Default "vu"}

\item{states_step_ratio}{put the ratio to change reservoir discretization in percent
0.01 to augment by 1\%}

\item{bellman_week}{Numeric of length 1. number of the week to plot
the correspondent Bellman values.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a \code{ggplot} object
}
\description{
Plot Bellman and Water values
}
