% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{plot_reward_variation}
\alias{plot_reward_variation}
\title{Plot the reward variation}
\usage{
plot_reward_variation(reward_base, week_id)
}
\arguments{
\item{reward_base}{A data.table contains the rewards.
Obtained using the function get_Reward()}

\item{week_id}{Numeric of length 1. number of the week to plot.}
}
\value{
a \code{ggplot} object
}
\description{
Plot the reward variation
}
