% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readReservoirLevels.R
\name{readReservoirLevels}
\alias{readReservoirLevels}
\title{Read Reservoir Levels}
\usage{
readReservoirLevels(
  area,
  timeStep = "weekly",
  byReservoirCapacity = TRUE,
  opts = antaresRead::simOptions(),
  force_final_level = F,
  final_level_egal_initial = F,
  final_level = NULL
)
}
\arguments{
\item{area}{An 'antares' area.}

\item{timeStep}{Resolution of the data to import: weekly (default, a linear interpolation is done on the data), monthly (original data).}

\item{byReservoirCapacity}{Multiply the result by the reservoir capacity (if available).}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{force_final_level}{Binary. Whether final level should be constrained}

\item{final_level_egal_initial}{Binary. Whether final level, if constrained, should be equal to initial level}

\item{final_level}{Final level (in percent between 0 and 100) if final level is constrained but different from initial level}
}
\value{
a data.table
}
\description{
Read Reservoir Levels
}
