% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readReservoirLevels.R
\name{readReservoirLevels}
\alias{readReservoirLevels}
\title{Read reservoir rule curves}
\usage{
readReservoirLevels(area, opts)
}
\arguments{
\item{area}{Character. The Antares area concerned by water values computation.}

\item{opts}{List of study parameters returned by the function \code{antaresRead::setSimulationPath(simulation="input")} in input mode.}
}
\value{
A \code{dplyr::tibble()} with 4 columns : \code{"timeId"} (index of the week), \code{"level_low"}, \code{"level_avg"} and \code{"level_high"}. Values are given for the end of the last hour of the end of the week.
}
\description{
Read reservoir low and high levels of hydro object of the specified area.
}
