% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readReservoirLevels.R
\name{readReservoirLevels}
\alias{readReservoirLevels}
\title{Read Reservoir Levels}
\usage{
readReservoirLevels(area, opts = antaresRead::simOptions())
}
\arguments{
\item{area}{An 'antares' area.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
a data.table
}
\description{
Read Reservoir Levels
}
