% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{remove_out}
\alias{remove_out}
\title{Remove outliers water Values}
\usage{
remove_out(results_dt, min = NULL, max = NULL, NAN = T)
}
\arguments{
\item{results_dt}{Output from \code{watervalues} or \code{Grid_Matrix}}

\item{min}{minimal accepted water value}

\item{max}{maximal accepted water value}

\item{NAN}{boolean. True to replace outlier values by 'NaN' and False to
replace them by 'min' and 'max' values}
}
\value{
a \code{data.table}
}
\description{
Remove outliers water Values
}
