% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_storage_functions.R
\name{resetHydroStorage}
\alias{resetHydroStorage}
\title{Reset to 0 the hydro storage time series, used in \code{setupWaterValuesSimulation}}
\usage{
resetHydroStorage(area, opts)
}
\arguments{
\item{area}{A valid Antares area.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Reset to 0 the hydro storage time series, used in \code{setupWaterValuesSimulation}
}
\note{
You can restore the original hydro storage time series with \code{restoreHydroStorage}.
}
\seealso{
\link{restoreHydroStorage}
}
\keyword{internal}
