% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_storage_functions.R
\name{resetHydroStorage}
\alias{resetHydroStorage}
\title{Reset to 0 the hydro storage time series}
\usage{
resetHydroStorage(area, path = NULL, opts = antaresRead::simOptions())
}
\arguments{
\item{area}{A valid Antares area.}

\item{path}{Optional, a path where to save the hydro storage file.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Reset to 0 the hydro storage time series
}
\note{
The function makes a copy of the original hydro storage time series,
 you can restore these with \code{restoreHydroStorage}.
}
\seealso{
\link{restoreHydroStorage}
}
