% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWaterValuesSimulation.R
\name{resetStudy}
\alias{resetStudy}
\title{Reset an Antares study. In case, there is a problem when executing \code{runWaterValuesSimulation()},
run this function to restore the study.}
\usage{
resetStudy(opts, area, pumping, backup)
}
\arguments{
\item{opts}{List of study parameters returned by the function \code{antaresRead::setSimulationPath(simulation="input")} in input mode.}

\item{area}{Character. The Antares area concerned by water values computation.}

\item{pumping}{Boolean. True to take into account the pumping capacity.}

\item{backup}{List with hydro_storage, load and misc_gen backups for area
generated by the function \code{getBackupData()}}
}
\description{
Reset an Antares study. In case, there is a problem when executing \code{runWaterValuesSimulation()},
run this function to restore the study.
}
\keyword{internal}
