% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_storage_functions.R
\name{restoreHydroStorage}
\alias{restoreHydroStorage}
\title{Restore the hydro storage time series, used in \code{runWaterValuesSimulation}}
\usage{
restoreHydroStorage(area, opts = antaresRead::simOptions(), data)
}
\arguments{
\item{area}{A valid Antares area.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{data}{Backup hydro storage matrix}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Restore the hydro storage time series, used in \code{runWaterValuesSimulation}
}
