% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Reward.R
\name{reward_offset}
\alias{reward_offset}
\title{Modify local reward to take into account overall cost of the simulation, mainly used in \code{Get_Reward}}
\usage{
reward_offset(
  opts,
  df_reward,
  u0 = c(),
  mcyears,
  district_cost = "water values district",
  fictive_areas = NULL,
  expansion = F
)
}
\arguments{
\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{df_reward}{data.table computed by the function \code{get_local_reward}}

\item{u0}{data.table {week,u} Constraint values per week used in the simulation, empty list if none}

\item{mcyears}{Vector of years used to evaluate rewards}

\item{district_cost}{Name of district used to evaluate overall cost}

\item{fictive_areas}{Vector of chr. Fictive areas used in simulation}

\item{expansion}{Binary. True if mode expansion was used to run simulations}
}
\value{
a data.table {mcYear,week,u,reward}
}
\description{
Modify local reward to take into account overall cost of the simulation, mainly used in \code{Get_Reward}
}
