% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWaterValuesSimulation.R
\name{runWaterValuesSimulation}
\alias{runWaterValuesSimulation}
\title{Run a simulation for calculating water values for a specific area}
\usage{
runWaterValuesSimulation(
  area,
  simulation_name = "weekly_water_amount_\%s",
  nb_disc_stock = 10,
  nb_mcyears = NULL,
  binding_constraint = "WeeklyWaterAmount",
  fictive_area = NULL,
  thermal_cluster = NULL,
  path_solver = NULL,
  wait = TRUE,
  show_output_on_console = FALSE,
  overwrite = FALSE,
  opts = antaresRead::simOptions(),
  shiny = F,
  otp_dest = NULL,
  file_name = NULL,
  ...
)
}
\arguments{
\item{area}{The area concerned by the simulation.}

\item{simulation_name}{The name of the simulation, \code{s} is a placeholder for the constraint value defined by \code{nb_disc_stock}.}

\item{nb_disc_stock}{Number of simulation to launch, a vector of energy constraint
will be created from 0 to the hydro storage maximum and of length this parameter.}

\item{nb_mcyears}{Number of Monte Carlo years to simulate.}

\item{binding_constraint}{Name of the binding constraint.}

\item{fictive_area}{Name of the fictive area to create, argument passed to \code{\link{setupWaterValuesSimulation}}.}

\item{thermal_cluster}{Name of the thermal cluster to create, argument passed to \code{\link{setupWaterValuesSimulation}}.}

\item{path_solver}{Character containing the Antares Solver path, argument passed to \code{\link[antaresEditObject]{runSimulation}}.}

\item{wait}{Argument passed to \code{\link[antaresEditObject]{runSimulation}}.}

\item{show_output_on_console}{Argument passed to \code{\link[antaresEditObject]{runSimulation}}.}

\item{overwrite}{If area or cluster already exists, should they be overwritten?}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Run a simulation for calculating water values for a specific area
}
\note{
This function have side effects on the Antares study used, a fictive area is created and a new district as well.
}
