% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWaterValuesSimulation.R
\name{runWaterValuesSimulation}
\alias{runWaterValuesSimulation}
\title{Run a simulation for calculating water values for a specific area}
\usage{
runWaterValuesSimulation(
  area = NULL,
  simulation_name = "wv_sim_\%s",
  nb_disc_stock = 10,
  mcyears = NULL,
  path_solver = NULL,
  show_output_on_console = FALSE,
  overwrite = FALSE,
  opts = antaresRead::simOptions(),
  file_name = NULL,
  pumping = T,
  efficiency,
  launch_simulations = NULL,
  constraint_values = NULL,
  expansion = T
)
}
\arguments{
\item{area}{The area concerned by the simulation.}

\item{simulation_name}{The name of the simulation, \code{s} is a placeholder for the constraint value defined by \code{nb_disc_stock}.}

\item{nb_disc_stock}{Number of simulation to launch, a vector of energy constraint
will be created from maximum pumping power to the hydro storage maximum and of length this parameter.}

\item{mcyears}{Number of Monte Carlo years to simulate or a vector of years indexes to launch.}

\item{path_solver}{Character containing the Antares Solver path, argument passed to \code{\link[antaresEditObject]{runSimulation}}.}

\item{show_output_on_console}{Argument passed to \code{\link[antaresEditObject]{runSimulation}}.}

\item{overwrite}{If area or cluster already exists, should they be overwritten?}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{file_name}{the Rdata file name.}

\item{pumping}{Boolean. True to take into account the pumping.}

\item{efficiency}{in [0,1]. efficient ratio of pumping.}

\item{launch_simulations}{Boolean. True to to run the simulations.}

\item{constraint_values}{constraint values to use to run simulations, generated by the function \code{\link{constraint_generator}}}

\item{expansion}{Binary. True if mode expansion was used to run simulations}
}
\description{
Run a simulation for calculating water values for a specific area
}
\note{
This function have side effects on the Antares study used, a fictive area is created and a new district as well.
}
