% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWaterValuesSimulation.R
\name{runWaterValuesSimulation}
\alias{runWaterValuesSimulation}
\title{Run a simulation for calculating water values for a specific area}
\usage{
runWaterValuesSimulation(
  area,
  simulation_name = "wv_sim_\%s",
  nb_disc_stock = 10,
  mcyears = NULL,
  binding_constraint = "weekly_water_amount",
  fictive_area = NULL,
  thermal_cluster = NULL,
  path_solver = NULL,
  show_output_on_console = FALSE,
  overwrite = FALSE,
  link_from = NULL,
  remove_areas = NULL,
  opts = antaresRead::simOptions(),
  otp_dest = NULL,
  file_name = NULL,
  pumping = F,
  efficiency = NULL,
  launch_simulations = T,
  constraint_values = NULL,
  expansion = T
)
}
\arguments{
\item{area}{area}

\item{simulation_name}{character}

\item{nb_disc_stock}{integer}

\item{mcyears}{list of Monte Carlo years}

\item{binding_constraint}{character}

\item{fictive_area}{Name of the fictive area to create, argument passed to \code{\link{setupWaterValuesSimulation}}.}

\item{thermal_cluster}{Name of the thermal cluster to create, argument passed to \code{\link{setupWaterValuesSimulation}}.}

\item{path_solver}{Character containing the Antares Solver path, argument passed to \code{\link[antaresEditObject]{runSimulation}}.}

\item{show_output_on_console}{Argument passed to \code{\link[antaresEditObject]{runSimulation}}.}

\item{overwrite}{If area or cluster already exists, should they be overwritten?}

\item{link_from}{area that will be linked to the created fictive area. If it's
\code{NULL} it will takes the area concerned by the simulation.}

\item{remove_areas}{Character vector of area(s) to remove from the created district.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{otp_dest}{the path in which the script save Rdata file.}

\item{file_name}{the Rdata file name.}

\item{pumping}{Boolean. True to take into account the pumping.}

\item{efficiency}{in [0,1]. efficient ratio of pumping.}

\item{launch_simulations}{Boolean. True to to run the simulations.}

\item{constraint_values}{constraint values to use to run simulations, generated by the function \code{\link{constraint_generator}}}

\item{expansion}{Binary. True if mode expansion was used to run simulations}
}
\description{
Run a simulation for calculating water values for a specific area
}
\note{
This function have side effects on the Antares study used, a fictive area is created and a new district as well.
}
