% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWaterValuesSimulation.R
\name{runWaterValuesSimulation}
\alias{runWaterValuesSimulation}
\title{Run Antares simulations in order to compute water values for a specific area}
\usage{
runWaterValuesSimulation(
  area,
  simulation_name = "wv_sim_\%s",
  nb_disc_stock = 10,
  mcyears = NULL,
  binding_constraint = "weekly_water_amount",
  fictive_area = NULL,
  thermal_cluster = NULL,
  path_solver = NULL,
  wait = TRUE,
  show_output_on_console = FALSE,
  overwrite = FALSE,
  link_from = NULL,
  remove_areas = NULL,
  opts = antaresRead::simOptions(),
  shiny = F,
  otp_dest = NULL,
  file_name = NULL,
  pumping = F,
  efficiency = NULL,
  launch_simulations = T,
  reset_hydro = T,
  constraint_values = NULL,
  expansion = T
)
}
\arguments{
\item{area}{Character. The Antares area concerned by water values computation.}

\item{simulation_name}{Character. The name of the simulation, \code{s} is a placeholder for the constraint value defined by \code{nb_disc_stock}.}

\item{nb_disc_stock}{Integer. Number of simulations to launch, \code{nb_disc_stock} different controls values
will be generated from maximum pumping power to maximum generating pumping with \code{constraint_generator()}. Each control values will be evaluated with a simulation.}

\item{mcyears}{Vector of integer. Monte Carlo years used to compute water values.}

\item{binding_constraint}{Character. Name of the binding constraint.}

\item{fictive_area}{Name of the fictive area to create, argument passed to \code{\link{setupWaterValuesSimulation}}.}

\item{thermal_cluster}{Name of the thermal cluster to create, argument passed to \code{\link{setupWaterValuesSimulation}}.}

\item{path_solver}{Character containing the Antares Solver path, argument passed to \code{\link[antaresEditObject]{runSimulation}}.}

\item{wait}{Argument passed to \code{\link[antaresEditObject]{runSimulation}}.}

\item{show_output_on_console}{Argument passed to \code{\link[antaresEditObject]{runSimulation}}.}

\item{overwrite}{If area or cluster already exists, should they be overwritten?}

\item{link_from}{area that will be linked to the created fictive area. If it's
\code{NULL} it will takes the area concerned by the simulation.}

\item{remove_areas}{Character vector of area(s) to remove from the created district.}

\item{opts}{List of study parameters returned by the function \code{antaresRead::setSimulationPath(simulation="input")} in input mode.}

\item{shiny}{Boolean. True to run the script in shiny mod.}

\item{otp_dest}{the path in which the script save Rdata file.}

\item{file_name}{the Rdata file name.}

\item{pumping}{Boolean. True to take into account the pumping capacity.}

\item{efficiency}{Double between 0 and 1. Pumping efficiency ratio. Get it with \code{getPumpEfficiency()}.}

\item{launch_simulations}{Boolean. True to to run the simulations.}

\item{reset_hydro}{Boolean. True to reset hydro inflow to 0 before the simulation.}

\item{constraint_values}{Constraint values to use to run simulations, generated by the function \code{\link{constraint_generator}} instead of parameter \code{nb_disc_stock}.}

\item{expansion}{Binary. True if mode expansion of Antares is used to run simulations, argument passed to \code{\link[antaresEditObject]{runSimulation}}.}
}
\value{
\item{simulation_names}{Vector of character. Names of simulations.}
\item{simulation_values}{A \code{dplyr::tibble()} generated by \code{constraint_generator()} describing controls for each simulation.}
\item{area}{Parameter \code{area}.}
\item{mc_years}{Parameter \code{mcyears}.}
\item{pumping}{Parameter \code{pumping}.}
\item{eff}{Parameter \code{efficiency}.}
\item{expansion}{Parameter \code{expansion}.}
\item{fictive_areas}{Parameter \code{fictive_areas}.}
}
\description{
\code{nb_disc_stock} simulations are run. Each simulation evaluates the cost of a specific control for each week and each scenario.
A control is the difference between initial level and final level on a week.
Results of simulations can be used to compute reward function with \code{get_Reward()} and then to compute water values with \code{Grid_Matrix()}.
}
\details{
To implement a specific control of the storage in Antares the study is modified : \itemize{
\item{The hydro object is deactivated by setting inflows to 0 and set \code{reservoir_management} to \code{FALSE}.}
\item{Two fictive areas representing pumping and generating are created and linked to \code{area}. The fictive area for pumping has a high enough load with unsupplied cost of 0. The fictive area for generating has a thermal cluster with marginal cost 0.}
\item{A weekly binding constraint is created imposing that the difference between the flow coming from generating fictive area and the flow going to pumping fictive area multiplied to the pumping efficiency is equal to the control.}
}
At the end of simulations, the study is reset.
}
\seealso{
\code{runWaterValuesSimulationMultiStock()}
}
