% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWaterValuesSimulation.R
\name{runWaterValuesSimulation}
\alias{runWaterValuesSimulation}
\title{Run Antares simulations in order to compute water values for a specific area}
\usage{
runWaterValuesSimulation(
  area = NULL,
  simulation_name = "wv_sim_\%s",
  nb_disc_stock = 10,
  mcyears = NULL,
  path_solver = NULL,
  show_output_on_console = FALSE,
  overwrite = FALSE,
  opts,
  file_name = NULL,
  pumping = TRUE,
  efficiency,
  launch_simulations = NULL,
  constraint_values = NULL,
  expansion = TRUE
)
}
\arguments{
\item{area}{Character. The Antares area concerned by water values computation.}

\item{simulation_name}{Character. The name of the simulation, \code{s} is a placeholder for the constraint value defined by \code{nb_disc_stock}.}

\item{nb_disc_stock}{Integer. Number of simulations to launch, \code{nb_disc_stock} different controls values
will be generated from maximum pumping power to maximum generating pumping with \code{constraint_generator()}. Each control values will be evaluated with a simulation.}

\item{mcyears}{Vector of integer. Monte Carlo years used to compute water values.}

\item{path_solver}{Character containing the Antares Solver path, argument passed to \code{\link[antaresEditObject]{runSimulation}}.}

\item{show_output_on_console}{Argument passed to \code{\link[antaresEditObject]{runSimulation}}.}

\item{overwrite}{If area or cluster already exists, should they be overwritten?}

\item{opts}{List of study parameters returned by the function \code{antaresRead::setSimulationPath(simulation="input")} in input mode.}

\item{file_name}{the Rdata file name.}

\item{pumping}{Boolean. True to take into account the pumping capacity.}

\item{efficiency}{Double between 0 and 1. Pumping efficiency ratio. Get it with \code{getPumpEfficiency()}.}

\item{launch_simulations}{Boolean or vector of booleans with the same length as \code{nb_disc_stock}. True to to run the simulations.}

\item{constraint_values}{Constraint values to use to run simulations, generated by the function \code{\link{constraint_generator}} instead of parameter \code{nb_disc_stock}.}

\item{expansion}{Binary. True if mode expansion (ie linear relaxation) of Antares is used to run simulations, argument passed to \code{\link[antaresEditObject]{runSimulation}}.
It is recommended to use mode expansion, it will be faster (only one iteration is done) and results will be smoother as the cost result will correspond to the linear relaxation of the problem.}
}
\value{
\item{simulation_names}{Vector of character. Names of simulations.}
\item{simulation_values}{A \code{dplyr::tibble()} generated by \code{constraint_generator()} describing controls for each simulation.}
\item{area}{Parameter \code{area}.}
\item{mcyears}{Parameter \code{mcyears}.}
\item{pumping}{Parameter \code{pumping}.}
\item{efficiency}{Parameter \code{efficiency}.}
\item{expansion}{Parameter \code{expansion}.}
\item{fictive_areas}{Parameter \code{fictive_areas}.}
}
\description{
\code{nb_disc_stock} simulations are run. Each simulation evaluates the cost of a specific control for each week and each scenario.
A control is the difference between initial level and final level on a week.
Results of simulations can be used to compute reward function with \code{get_Reward()} and then to compute water values with \code{Grid_Matrix()}.
See \code{vignette("computation_watervalues")} for more details.
}
\details{
To implement a specific control of the storage in Antares the study is modified : \itemize{
\item{The hydro object is deactivated by setting inflows to 0 and set \code{reservoir_management} to \code{FALSE}.}
\item{Two fictive areas representing pumping and generating are created and linked to \code{area}. The fictive area for pumping has a high enough load with unsupplied cost of 0. The fictive area for generating has a thermal cluster with marginal cost 0.}
\item{A weekly binding constraint is created imposing that the difference between the flow coming from generating fictive area and the flow going to pumping fictive area multiplied to the pumping efficiency is equal to the control.}
}
At the end of simulations, the study is reset.
}
\seealso{
\code{runWaterValuesSimulationMultiStock()}
}
