% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWaterValuesSimulation.R
\name{runWaterValuesSimulationMultiStock}
\alias{runWaterValuesSimulationMultiStock}
\title{Run a simulation for calculating water values for a specific area}
\usage{
runWaterValuesSimulationMultiStock(
  list_areas,
  list_pumping,
  list_efficiency,
  simulation_name = "wv_sim_\%s",
  mcyears = NULL,
  path_solver = NULL,
  show_output_on_console = FALSE,
  overwrite = FALSE,
  opts = antaresRead::simOptions(),
  file_name = NULL,
  launch_simulations = NULL,
  constraint_values = NULL,
  expansion = T
)
}
\arguments{
\item{list_areas}{List of areas concerned by the simulation.}

\item{list_pumping}{List of bools to tell if pumping is available in areas}

\item{list_efficiency}{List of pumping efficiency}

\item{simulation_name}{character}

\item{mcyears}{Number of Monte Carlo years to simulate or a vector of years indexes to launch.}

\item{path_solver}{Character containing the Antares Solver path, argument passed to \code{\link[antaresEditObject]{runSimulation}}.}

\item{show_output_on_console}{Argument passed to \code{\link[antaresEditObject]{runSimulation}}.}

\item{overwrite}{If area or cluster already exists, should they be overwritten?}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{file_name}{the Rdata file name.}

\item{launch_simulations}{Boolean. True to to run the simulations.}

\item{constraint_values}{constraint values to use to run simulations, generated by the function \code{\link{constraint_generator}}}

\item{expansion}{Binary. True if mode expansion was used to run simulations}
}
\description{
Run a simulation for calculating water values for a specific area
}
\note{
This function have side effects on the Antares study used, a fictive area is created and a new district as well.
}
