% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupWaterValuesSimulation.R
\name{setupWaterValuesSimulation}
\alias{setupWaterValuesSimulation}
\title{Setup a simulation before running it for calculating Water Values,
used in \code{runWaterValuesSimulation}}
\usage{
setupWaterValuesSimulation(
  area,
  fictive_area_name = paste0("watervalue_", area),
  thermal_cluster = "water_value_cluster",
  overwrite = FALSE,
  remove_areas = NULL,
  opts = antaresRead::simOptions(),
  reset_hydro = T,
  link_from = NULL,
  pumping = F,
  max_load = 1e+08,
  ...
)
}
\arguments{
\item{area}{The area concerned by the simulation.}

\item{fictive_area_name}{Name of the fictive area to create.}

\item{thermal_cluster}{Name of the thermal cluster to create.}

\item{overwrite}{If area or cluster already exists, overwrite them ?}

\item{remove_areas}{Character vector of area(s) to remove from the created district.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{reset_hydro}{Boolean. True to reset hydro inflow to 0 before the simulation.}

\item{link_from}{area that will be linked to the created fictive area. If it's
\code{NULL} it will takes the area concerned by the simulation.}

\item{pumping}{Boolean. True to take into account the pumping.}

\item{max_load}{the maximum load to put it in the fictive areas.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The result of antaresRead::simOptions
}
\description{
Setup a simulation before running it for calculating Water Values,
used in \code{runWaterValuesSimulation}
}
\keyword{internal}
