% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupWaterValuesSimulation.R
\name{setupWaterValuesSimulation}
\alias{setupWaterValuesSimulation}
\title{Setup a simulation before running it for calculating Water Values,
used in \code{runWaterValuesSimulation}}
\usage{
setupWaterValuesSimulation(
  area,
  overwrite = FALSE,
  opts = antaresRead::simOptions(),
  pumping = F,
  efficiency,
  backup
)
}
\arguments{
\item{area}{The area concerned by the simulation.}

\item{overwrite}{If area or cluster already exists, overwrite them ?}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{pumping}{Boolean. True to take into account the pumping.}

\item{efficiency}{Double. Pumping efficiency}

\item{backup}{List with hydro_storage, load and misc_gen backups for area
generated by the function \code{getBackupData}}
}
\value{
The result of antaresRead::simOptions
}
\description{
Setup a simulation before running it for calculating Water Values,
used in \code{runWaterValuesSimulation}
}
\keyword{internal}
