% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupWaterValuesSimulation.R
\name{setupWaterValuesSimulation}
\alias{setupWaterValuesSimulation}
\title{Setup a simulation before running it for calculating Water Values}
\usage{
setupWaterValuesSimulation(
  area,
  fictive_area = paste0("watervalue_", area),
  thermal_cluster = "WaterValueCluster",
  overwrite = FALSE,
  remove_areas = NULL,
  opts = antaresRead::simOptions(),
  ...
)
}
\arguments{
\item{area}{The area concerned by the simulation.}

\item{fictive_area}{Name of the fictive area to create.}

\item{thermal_cluster}{Name of the thermal cluster to create.}

\item{overwrite}{If area or cluster already exists, overwrite them ?}

\item{remove_areas}{Character vector of area(s) to remove from the created district.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
The result of antaresRead::simOptions
}
\description{
Setup a simulation before running it for calculating Water Values
}
