% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{shiny_Grid_matrix}
\alias{shiny_Grid_matrix}
\title{Open watervalues Calculator in APP Web}
\usage{
shiny_Grid_matrix(simulation_res = NULL, study_path, ...)
}
\arguments{
\item{simulation_res}{List of simulation results returned by the function
\code{watervalues::runWaterValuesSimulation}}

\item{study_path}{the path of the Antares study}
}
\description{
Open watervalues Calculator in APP Web
}
