% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{shiny_water_values}
\alias{shiny_water_values}
\title{Open web interface for computing water values}
\usage{
shiny_water_values(opts, silent = FALSE)
}
\arguments{
\item{opts}{List of study parameters returned by the function \code{antaresRead::setSimulationPath(simulation="input")} in input mode.}

\item{silent}{Binary. \code{TRUE} to suppress warnings.}
}
\description{
Open web interface for computing water values. Select the study with parameter \code{opts}.
}
