% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{shiny_water_values}
\alias{shiny_water_values}
\title{Open web interface for computing water values}
\usage{
shiny_water_values(opts = NULL, simulation_res = NULL, silent = F, ...)
}
\arguments{
\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{simulation_res}{List of simulation results returned by the function
\code{watervalues::runWaterValuesSimulation}}

\item{silent}{Boolean. TRUE to suppress warnings.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Open web interface for computing water values
}
