% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{to_Antares_Format}
\alias{to_Antares_Format}
\title{Convert water values to Antares format}
\usage{
to_Antares_Format(data, constant = TRUE)
}
\arguments{
\item{data}{A \code{dplyr::tibble()} representing water values, output \code{aggregated_results} generated by \code{Grid_Matrix()}.}

\item{constant}{Binary. Generate daily constant values by week. \code{FALSE} to do interpolation.}
}
\value{
A 365*101 numeric matrix
}
\description{
This function converts water values generated by \code{Grid_Matrix()}
to the format expected by Antares: a 365*101 matrix, where
the rows are the 365 days of the year and the columns are round percentage values
ranging from 0 to 100 assessing the reservoir level.
Use \code{antaresEditObject::writeWaterValues()} to write water values.
}
\note{
Please use \code{to_Antares_Format_bis()} if you are using \code{hydro-pricing-mode=accurate} in Antares.
}
\seealso{
\code{to_Antares_Format_bis()}
}
