% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watervaluesViz.R
\name{waterValuesViz}
\alias{waterValuesViz}
\title{Plot water values}
\usage{
waterValuesViz(Data, filter_penalties = FALSE)
}
\arguments{
\item{Data}{Water values result. Output \code{aggregated_results} from \code{Grid_Matrix()}.}

\item{filter_penalties}{Binary. If \code{TRUE}, show only water values inside rule curves.}
}
\value{
A \code{ggplot2::ggplot()} object.
}
\description{
Plot water values for all weeks and all levels of stock.
}
