% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watervaluesViz.R
\name{waterValuesViz}
\alias{waterValuesViz}
\title{Visualize mean grid layer}
\usage{
waterValuesViz(Data, filter_penalties = FALSE)
}
\arguments{
\item{Data}{Output from \code{Grid_Matrix}.}

\item{filtre_ratio}{in [1,0]. Define the percent to keep from water values
eliminating the rest (extreme negatives and positives).}

\item{show_negative}{Boolean. true to show negative water values and
False to delete them from the graph.}
}
\value{
a \code{ggplot} object
}
\description{
Visualize mean grid layer
}
