% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watervaluesViz.R
\name{waterValuesViz}
\alias{waterValuesViz}
\title{Visualize mean grid layer}
\usage{
waterValuesViz(Data, filter_penalties = FALSE)
}
\arguments{
\item{Data}{Output from \code{Grid_Matrix}.}

\item{filter_penalties}{Binary paramater. If T, show only watervalues inside rule curves}
}
\value{
a \code{ggplot} object
}
\description{
Visualize mean grid layer
}
